First song choice model

This model was estimated on only the data for the first song of each show.

ml.Repeatr4_fs

Format

An object of class mlogit of length 17.

Details

Songs that were never ever played as the first song of a Fugazi song were excluded, by necessity.

The utility formula was as follows:

choice ~ yearsold_1 + yearsold_2 + yearsold_3 + song2 + ... + song92

Where yearsold_3 = yearsold_3 + yearsold_4 + yearsold_5 + yearsold_6 + yearsold_7 + yearsold_8

Examples

  ml.Repeatr4_fs
#> 
#> Call:
#> mlogit(formula = choice ~ yearsold_1 + yearsold_2 + yearsold_3,     data = Repeatr4_fs, method = "nr")
#> 
#> Coefficients:
#>  (Intercept):2   (Intercept):3   (Intercept):4   (Intercept):5   (Intercept):7  
#>     3.2687e+00      2.5324e+00      1.2269e+00      2.7515e+00      1.6154e+00  
#>  (Intercept):8   (Intercept):9  (Intercept):10  (Intercept):11  (Intercept):12  
#>     2.3842e-01      3.0595e+00     -2.7405e-01      1.3071e+00      1.3440e+00  
#> (Intercept):13  (Intercept):14  (Intercept):15  (Intercept):17  (Intercept):18  
#>    -3.6511e-01      5.0696e-01     -4.0503e-01      1.5404e+00      2.4571e+00  
#> (Intercept):19  (Intercept):20  (Intercept):21  (Intercept):22  (Intercept):23  
#>     3.2403e+00      1.7881e+00      3.0175e-01      2.5290e+00      1.6201e+00  
#> (Intercept):24  (Intercept):25  (Intercept):26  (Intercept):27  (Intercept):28  
#>     2.5907e+00      5.1462e-01      5.0696e-01     -1.5467e-01      1.3545e+00  
#> (Intercept):29  (Intercept):30  (Intercept):31  (Intercept):32  (Intercept):33  
#>     1.6405e+00      3.4204e+00      1.4508e+00      1.9174e+00      9.1827e-02  
#> (Intercept):34  (Intercept):35  (Intercept):37  (Intercept):38  (Intercept):40  
#>    -3.8281e-01     -2.9958e-01      4.1254e+00      9.2227e-01      7.0536e-01  
#> (Intercept):41  (Intercept):42  (Intercept):43  (Intercept):45  (Intercept):46  
#>     2.5882e+00      5.5643e+00     -4.0503e-01      3.6949e+00      3.4049e+00  
#> (Intercept):47  (Intercept):48  (Intercept):50  (Intercept):51  (Intercept):52  
#>     1.8205e-01      1.7258e-01      6.7955e-01     -3.5623e-01      1.6098e+00  
#> (Intercept):53  (Intercept):54  (Intercept):55  (Intercept):56  (Intercept):57  
#>    -1.0845e+00      2.5835e-01     -1.3600e-02      1.4700e+00     -1.0491e+00  
#> (Intercept):58  (Intercept):60  (Intercept):61  (Intercept):62  (Intercept):64  
#>     3.4864e+00      1.8578e-01      1.6146e+00      1.2044e+00      4.6966e+00  
#> (Intercept):66  (Intercept):68  (Intercept):69  (Intercept):70  (Intercept):71  
#>     2.2092e-15      2.0781e+00      1.2633e+00      2.5732e+00      4.2884e+00  
#> (Intercept):72  (Intercept):73  (Intercept):74  (Intercept):75  (Intercept):78  
#>     1.2280e+00      5.9054e+00      3.2873e+00      2.5732e+00      1.6807e+00  
#> (Intercept):80  (Intercept):81  (Intercept):84  (Intercept):85  (Intercept):86  
#>     3.5494e+00     -1.0986e+00     -1.0986e+00      2.9755e+00      3.0009e+00  
#> (Intercept):89  (Intercept):91      yearsold_1      yearsold_2      yearsold_3  
#>    -1.0986e+00     -4.8884e-01     -3.8494e-01     -3.7746e-01     -3.1163e-01  
#>