Other song choice model

This model was estimated on only the data for the other songs of each show, that is all the songs after the first song.

ml.Repeatr4_os

Format

An object of class mlogit of length 17.

Details

The utility formula was as follows:

choice ~ choice ~ choice ~ yearsold_1 + yearsold_2 + yearsold_3 + yearsold_4 + yearsold_5 + yearsold_6 + yearsold_7 + yearsold_8 + song2 + ... + song92

Examples

  ml.Repeatr4_os
#> 
#> Call:
#> mlogit(formula = choice ~ yearsold_1 + yearsold_2 + yearsold_3 +     yearsold_4 + yearsold_5 + yearsold_6 + yearsold_7 + yearsold_8,     data = Repeatr4_os, method = "nr")
#> 
#> Coefficients:
#>  (Intercept):2   (Intercept):3   (Intercept):4   (Intercept):5   (Intercept):6  
#>       0.370037       -0.197652       -0.481492       -0.292538       -0.377837  
#>  (Intercept):7   (Intercept):8   (Intercept):9  (Intercept):10  (Intercept):11  
#>      -0.197771       -0.857391       -0.544908       -0.926293       -0.620015  
#> (Intercept):12  (Intercept):13  (Intercept):14  (Intercept):15  (Intercept):16  
#>      -0.814923       -0.786519       -0.243449       -0.891986       -1.011234  
#> (Intercept):17  (Intercept):18  (Intercept):19  (Intercept):20  (Intercept):21  
#>      -0.806446        0.435333       -0.513569       -0.933299       -0.990133  
#> (Intercept):22  (Intercept):23  (Intercept):24  (Intercept):25  (Intercept):26  
#>      -1.094891       -1.343487        0.243254       -0.693010       -0.784599  
#> (Intercept):27  (Intercept):28  (Intercept):29  (Intercept):30  (Intercept):31  
#>      -1.174022       -0.507313       -0.030867       -0.756825       -1.087096  
#> (Intercept):32  (Intercept):33  (Intercept):34  (Intercept):35  (Intercept):36  
#>      -0.635608       -1.287327       -1.376437       -1.363352       -0.592163  
#> (Intercept):37  (Intercept):38  (Intercept):39  (Intercept):40  (Intercept):41  
#>      -0.595978       -1.480763       -1.419833       -1.792780       -1.482574  
#> (Intercept):42  (Intercept):43  (Intercept):44  (Intercept):45  (Intercept):46  
#>      -0.034677       -1.722341        0.120467       -2.679545       -0.390457  
#> (Intercept):47  (Intercept):48  (Intercept):49  (Intercept):50  (Intercept):51  
#>      -1.624262       -1.585646       -0.071062       -1.320462       -1.690270  
#> (Intercept):52  (Intercept):53  (Intercept):54  (Intercept):55  (Intercept):56  
#>      -1.034961       -1.951933       -1.344544       -1.420619        0.020857  
#> (Intercept):57  (Intercept):58  (Intercept):59  (Intercept):60  (Intercept):61  
#>      -2.088502       -3.418639       -0.607576       -1.900219       -1.400888  
#> (Intercept):62  (Intercept):63  (Intercept):64  (Intercept):65  (Intercept):66  
#>      -1.314848       -0.649176       -0.349345       -0.672135       -2.333579  
#> (Intercept):67  (Intercept):68  (Intercept):69  (Intercept):70  (Intercept):71  
#>      -0.963841       -0.573043       -1.441028       -0.546586       -0.016249  
#> (Intercept):72  (Intercept):73  (Intercept):74  (Intercept):75  (Intercept):76  
#>      -1.676536       -1.781571       -5.283173       -1.126573       -2.434047  
#> (Intercept):77  (Intercept):78  (Intercept):79  (Intercept):80  (Intercept):81  
#>      -0.534778       -2.763608       -1.974286       -2.373887       -3.616297  
#> (Intercept):82  (Intercept):83  (Intercept):84  (Intercept):85  (Intercept):86  
#>      -0.963607       -3.199963       -3.763027       -0.939768       -3.946540  
#> (Intercept):87  (Intercept):88  (Intercept):89  (Intercept):90  (Intercept):91  
#>      -0.851986       -1.393190       -4.311272       -2.568696       -5.173675  
#> (Intercept):92      yearsold_1      yearsold_2      yearsold_3      yearsold_4  
#>      -3.516750        0.250158       -0.065885       -0.315286       -0.644406  
#>     yearsold_5      yearsold_6      yearsold_7      yearsold_8  
#>      -0.813249       -0.903761       -1.000783       -0.848459  
#>