Estimated coefficients and related statistics from the model ml.Repeatr4_fs

First song choice model

results_ml_Repeatr4_fs

Format

dataframe with one row for each coefficient in the model.

Estimate

The coefficient value

Std. Error

The standard error of the coefficient

z-value

The z-value of the coefficient

Pr(>|z|)

The P value of the coefficient

Details

This model was estimated with mlogit on only the data for the first song of each show.

Songs that were never ever played as the first song of a Fugazi song were excluded, by necessity.

The utility formula was as follows:

choice ~ yearsold_1 + yearsold_2 + yearsold_3 + song2 + ... + song92

Where yearsold_3 = yearsold_3 + yearsold_4 + yearsold_5 + yearsold_6 + yearsold_7 + yearsold_8

Examples

  results_ml_Repeatr4_fs
#>                     Estimate Std. Error       z-value     Pr(>|z|)
#> (Intercept):2   3.268691e+00  0.5980329  5.465737e+00 4.609872e-08
#> (Intercept):3   2.532352e+00  0.6102348  4.149799e+00 3.327677e-05
#> (Intercept):4   1.226906e+00  0.6794937  1.805618e+00 7.097801e-02
#> (Intercept):5   2.751535e+00  0.5954913  4.620614e+00 3.826069e-06
#> (Intercept):7   1.615417e+00  0.6480507  2.492732e+00 1.267645e-02
#> (Intercept):8   2.384199e-01  0.8193347  2.909921e-01 7.710574e-01
#> (Intercept):9   3.059488e+00  0.5972535  5.122595e+00 3.013592e-07
#> (Intercept):10 -2.740519e-01  0.9140362 -2.998261e-01 7.643098e-01
#> (Intercept):11  1.307075e+00  0.6513760  2.006637e+00 4.478832e-02
#> (Intercept):12  1.344046e+00  0.6692923  2.008160e+00 4.462633e-02
#> (Intercept):13 -3.651063e-01  0.9133280 -3.997537e-01 6.893379e-01
#> (Intercept):14  5.069614e-01  0.9207442  5.505997e-01 5.819081e-01
#> (Intercept):15 -4.050314e-01  0.9131241 -4.435667e-01 6.573559e-01
#> (Intercept):17  1.540445e+00  0.6362090  2.421288e+00 1.546563e-02
#> (Intercept):18  2.457127e+00  0.7132549  3.444949e+00 5.711666e-04
#> (Intercept):19  3.240276e+00  0.6122992  5.291982e+00 1.209981e-07
#> (Intercept):20  1.788078e+00  0.6462628  2.766798e+00 5.660989e-03
#> (Intercept):21  3.017529e-01  0.7638714  3.950310e-01 6.928200e-01
#> (Intercept):22  2.529024e+00  0.6179488  4.092611e+00 4.265430e-05
#> (Intercept):23  1.620088e+00  0.6478210  2.500826e+00 1.239040e-02
#> (Intercept):24  2.590658e+00  0.7006250  3.697639e+00 2.176141e-04
#> (Intercept):25  5.146173e-01  0.9208345  5.588597e-01 5.762575e-01
#> (Intercept):26  5.069614e-01  0.9207442  5.505997e-01 5.819081e-01
#> (Intercept):27 -1.546688e-01  0.9161171 -1.688308e-01 8.659297e-01
#> (Intercept):28  1.354474e+00  0.7783012  1.740296e+00 8.180713e-02
#> (Intercept):29  1.640464e+00  0.8343803  1.966086e+00 4.928865e-02
#> (Intercept):30  3.420387e+00  0.6210641  5.507301e+00 3.643774e-08
#> (Intercept):31  1.450754e+00  0.7413073  1.957021e+00 5.034505e-02
#> (Intercept):32  1.917409e+00  0.7234933  2.650210e+00 8.044175e-03
#> (Intercept):33  9.182658e-02  0.8174362  1.123349e-01 9.105579e-01
#> (Intercept):34 -3.828085e-01  0.9132415 -4.191756e-01 6.750878e-01
#> (Intercept):35 -2.995801e-01  0.9138010 -3.278396e-01 7.430330e-01
#> (Intercept):37  4.125370e+00  0.6250079  6.600509e+00 4.097478e-11
#> (Intercept):38  9.222695e-01  0.7094385  1.299999e+00 1.936013e-01
#> (Intercept):40  7.053618e-01  0.7071794  9.974298e-01 3.185559e-01
#> (Intercept):41  2.588209e+00  0.6316388  4.097610e+00 4.174387e-05
#> (Intercept):42  5.564299e+00  0.6243372  8.912330e+00 0.000000e+00
#> (Intercept):43 -4.050314e-01  0.9131241 -4.435667e-01 6.573559e-01
#> (Intercept):45  3.694858e+00  0.5891120  6.271911e+00 3.566440e-10
#> (Intercept):46  3.404864e+00  0.7389119  4.607944e+00 4.066707e-06
#> (Intercept):47  1.820463e-01  0.9188884  1.981158e-01 8.429545e-01
#> (Intercept):48  1.725799e-01  0.9188122  1.878294e-01 8.510104e-01
#> (Intercept):50  6.795468e-01  0.9238304  7.355753e-01 4.619892e-01
#> (Intercept):51 -3.562298e-01  1.1603172 -3.070107e-01 7.588352e-01
#> (Intercept):52  1.609829e+00  0.8361870  1.925202e+00 5.420407e-02
#> (Intercept):53 -1.084541e+00  1.1547725 -9.391819e-01 3.476373e-01
#> (Intercept):54  2.583543e-01  1.1650368  2.217564e-01 8.245035e-01
#> (Intercept):55 -1.360042e-02  1.1633841 -1.169039e-02 9.906726e-01
#> (Intercept):56  1.469955e+00  1.1718064  1.254435e+00 2.096841e-01
#> (Intercept):57 -1.049136e+00  1.1551504 -9.082243e-01 3.637597e-01
#> (Intercept):58  3.486355e+00  0.5861206  5.948187e+00 2.711288e-09
#> (Intercept):60  1.857830e-01  0.9189116  2.021772e-01 8.397782e-01
#> (Intercept):61  1.614623e+00  0.8361722  1.930969e+00 5.348687e-02
#> (Intercept):62  1.204364e+00  0.9305243  1.294285e+00 1.955669e-01
#> (Intercept):64  4.696585e+00  0.6824409  6.882039e+00 5.900169e-12
#> (Intercept):66  2.209221e-15  0.8164966  2.705732e-15 1.000000e+00
#> (Intercept):68  2.078129e+00  1.1828747  1.756846e+00 7.894403e-02
#> (Intercept):69  1.263318e+00  1.1706832  1.079129e+00 2.805304e-01
#> (Intercept):70  2.573163e+00  1.1886579  2.164763e+00 3.040584e-02
#> (Intercept):71  4.288393e+00  0.8267788  5.186868e+00 2.138599e-07
#> (Intercept):72  1.227965e+00  1.1699039  1.049629e+00 2.938889e-01
#> (Intercept):73  5.905367e+00  0.6697925  8.816712e+00 0.000000e+00
#> (Intercept):74  3.287319e+00  0.6070848  5.414925e+00 6.131454e-08
#> (Intercept):75  2.573163e+00  1.1886579  2.164763e+00 3.040584e-02
#> (Intercept):78  1.680653e+00  0.8343704  2.014277e+00 4.398051e-02
#> (Intercept):80  3.549396e+00  0.7057834  5.029016e+00 4.930036e-07
#> (Intercept):81 -1.098612e+00  1.1547005 -9.514262e-01 3.413881e-01
#> (Intercept):84 -1.098612e+00  1.1547005 -9.514262e-01 3.413881e-01
#> (Intercept):85  2.975504e+00  1.2009460  2.477633e+00 1.322570e-02
#> (Intercept):86  3.000917e+00  0.6702119  4.477565e+00 7.549937e-06
#> (Intercept):89 -1.098612e+00  1.1547005 -9.514262e-01 3.413881e-01
#> (Intercept):91 -4.888370e-01  1.1594981 -4.215936e-01 6.733217e-01
#> yearsold_1     -3.849420e-01  0.1213225 -3.172882e+00 1.509339e-03
#> yearsold_2     -3.774616e-01  0.1289066 -2.928180e+00 3.409530e-03
#> yearsold_3     -3.116347e-01  0.1557674 -2.000641e+00 4.543109e-02