Estimated coefficients and related statistics from the model ml.Repeatr4_os

Other song choice model

results_ml_Repeatr4_os

Format

dataframe with one row for each coefficient in the model.

Estimate

The coefficient value

Std. Error

The standard error of the coefficient

z-value

The z-value of the coefficient

Pr(>|z|)

The P value of the coefficient

Details

This model was estimated with mlogit on only the data for the other songs of each show, that is all the songs after the first song.

The utility formula was as follows:

choice ~ choice ~ choice ~ yearsold_1 + yearsold_2 + yearsold_3 + yearsold_4 + yearsold_5 + yearsold_6 + yearsold_7 + yearsold_8 + song2 + ... + song92

Examples

  results_ml_Repeatr4_os
#>                   Estimate Std. Error     z-value     Pr(>|z|)
#> (Intercept):2   0.37003687 0.06842796   5.4076851 6.384453e-08
#> (Intercept):3  -0.19765159 0.06635859  -2.9785383 2.896268e-03
#> (Intercept):4  -0.48149205 0.06484337  -7.4254626 1.123546e-13
#> (Intercept):5  -0.29253791 0.06460521  -4.5280853 5.952056e-06
#> (Intercept):6  -0.37783668 0.06537678  -5.7793712 7.498035e-09
#> (Intercept):7  -0.19777099 0.06817246  -2.9010394 3.719271e-03
#> (Intercept):8  -0.85739057 0.06947475 -12.3410387 0.000000e+00
#> (Intercept):9  -0.54490774 0.07136124  -7.6359068 2.242651e-14
#> (Intercept):10 -0.92629286 0.06845513 -13.5313861 0.000000e+00
#> (Intercept):11 -0.62001467 0.06804241  -9.1121803 0.000000e+00
#> (Intercept):12 -0.81492340 0.07041634 -11.5729304 0.000000e+00
#> (Intercept):13 -0.78651912 0.06768721 -11.6199081 0.000000e+00
#> (Intercept):14 -0.24344925 0.08735157  -2.7870050 5.319767e-03
#> (Intercept):15 -0.89198577 0.06909282 -12.9099639 0.000000e+00
#> (Intercept):16 -1.01123383 0.07255230 -13.9379979 0.000000e+00
#> (Intercept):17 -0.80644614 0.07121458 -11.3241716 0.000000e+00
#> (Intercept):18  0.43533308 0.11833797   3.6787269 2.344011e-04
#> (Intercept):19 -0.51356892 0.08940379  -5.7443755 9.226087e-09
#> (Intercept):20 -0.93329909 0.08041952 -11.6053806 0.000000e+00
#> (Intercept):21 -0.99013282 0.07391069 -13.3963413 0.000000e+00
#> (Intercept):22 -1.09489079 0.08397954 -13.0375898 0.000000e+00
#> (Intercept):23 -1.34348723 0.07681093 -17.4908342 0.000000e+00
#> (Intercept):24  0.24325419 0.12000633   2.0270113 4.266126e-02
#> (Intercept):25 -0.69300974 0.09197636  -7.5346503 4.907186e-14
#> (Intercept):26 -0.78459943 0.09204075  -8.5244793 0.000000e+00
#> (Intercept):27 -1.17402158 0.08022645 -14.6338473 0.000000e+00
#> (Intercept):28 -0.50731251 0.10143543  -5.0013344 5.693485e-07
#> (Intercept):29 -0.03086726 0.11635654  -0.2652817 7.907925e-01
#> (Intercept):30 -0.75682450 0.10274544  -7.3660156 1.758593e-13
#> (Intercept):31 -1.08709617 0.09655440 -11.2588980 0.000000e+00
#> (Intercept):32 -0.63560797 0.10520475  -6.0416281 1.525668e-09
#> (Intercept):33 -1.28732685 0.08028785 -16.0338937 0.000000e+00
#> (Intercept):34 -1.37643717 0.08104975 -16.9826198 0.000000e+00
#> (Intercept):35 -1.36335207 0.08177883 -16.6712096 0.000000e+00
#> (Intercept):36 -0.59216340 0.12139216  -4.8781024 1.071113e-06
#> (Intercept):37 -0.59597845 0.12596057  -4.7314682 2.229018e-06
#> (Intercept):38 -1.48076257 0.08773975 -16.8767579 0.000000e+00
#> (Intercept):39 -1.41983257 0.09361693 -15.1664082 0.000000e+00
#> (Intercept):40 -1.79277985 0.08822803 -20.3198454 0.000000e+00
#> (Intercept):41 -1.48257411 0.10377650 -14.2862214 0.000000e+00
#> (Intercept):42 -0.03467714 0.15624993  -0.2219338 8.243654e-01
#> (Intercept):43 -1.72234106 0.08976481 -19.1872626 0.000000e+00
#> (Intercept):44  0.12046688 0.15386425   0.7829427 4.336608e-01
#> (Intercept):45 -2.67954516 0.13729237 -19.5170732 0.000000e+00
#> (Intercept):46 -0.39045717 0.15385823  -2.5377724 1.115605e-02
#> (Intercept):47 -1.62426204 0.10074638 -16.1222862 0.000000e+00
#> (Intercept):48 -1.58564583 0.10159933 -15.6068530 0.000000e+00
#> (Intercept):49 -0.07106156 0.15434105  -0.4604191 6.452155e-01
#> (Intercept):50 -1.32046152 0.11490429 -11.4918384 0.000000e+00
#> (Intercept):51 -1.69027021 0.10732905 -15.7484870 0.000000e+00
#> (Intercept):52 -1.03496070 0.13534345  -7.6469210 2.065015e-14
#> (Intercept):53 -1.95193310 0.09628030 -20.2734425 0.000000e+00
#> (Intercept):54 -1.34454446 0.12196869 -11.0236854 0.000000e+00
#> (Intercept):55 -1.42061938 0.11616487 -12.2293374 0.000000e+00
#> (Intercept):56  0.02085736 0.15055672   0.1385349 8.898177e-01
#> (Intercept):57 -2.08850169 0.09737669 -21.4476564 0.000000e+00
#> (Intercept):58 -3.41863915 0.18341683 -18.6386342 0.000000e+00
#> (Intercept):59 -0.60757569 0.15958613  -3.8071960 1.405513e-04
#> (Intercept):60 -1.90021921 0.11156706 -17.0320803 0.000000e+00
#> (Intercept):61 -1.40088799 0.14095394  -9.9386226 0.000000e+00
#> (Intercept):62 -1.31484846 0.14088054  -9.3330736 0.000000e+00
#> (Intercept):63 -0.64917583 0.15548814  -4.1750827 2.978778e-05
#> (Intercept):64 -0.34934480 0.17961644  -1.9449489 5.178113e-02
#> (Intercept):65 -0.67213547 0.15697761  -4.2817284 1.854472e-05
#> (Intercept):66 -2.33357876 0.11347132 -20.5653624 0.000000e+00
#> (Intercept):67 -0.96384086 0.16242818  -5.9339510 2.957301e-09
#> (Intercept):68 -0.57304306 0.18691141  -3.0658538 2.170494e-03
#> (Intercept):69 -1.44102780 0.16982748  -8.4852450 0.000000e+00
#> (Intercept):70 -0.54658597 0.19680361  -2.7773168 5.480973e-03
#> (Intercept):71 -0.01624878 0.21329088  -0.0761813 9.392749e-01
#> (Intercept):72 -1.67653553 0.18297068  -9.1628643 0.000000e+00
#> (Intercept):73 -1.78157124 0.28110637  -6.3377121 2.332019e-10
#> (Intercept):74 -5.28317252 0.50364820 -10.4898072 0.000000e+00
#> (Intercept):75 -1.12657305 0.21767556  -5.1754686 2.273396e-07
#> (Intercept):76 -2.43404716 0.19066514 -12.7660840 0.000000e+00
#> (Intercept):77 -0.53477833 0.23953403  -2.2325777 2.557681e-02
#> (Intercept):78 -2.76360845 0.20508986 -13.4751104 0.000000e+00
#> (Intercept):79 -1.97428557 0.22390354  -8.8175720 0.000000e+00
#> (Intercept):80 -2.37388692 0.23805140  -9.9721611 0.000000e+00
#> (Intercept):81 -3.61629720 0.19176404 -18.8580566 0.000000e+00
#> (Intercept):82 -0.96360674 0.24570334  -3.9218300 8.787900e-05
#> (Intercept):83 -3.19996310 0.21085346 -15.1762416 0.000000e+00
#> (Intercept):84 -3.76302676 0.20845741 -18.0517773 0.000000e+00
#> (Intercept):85 -0.93976817 0.26900555  -3.4934899 4.767511e-04
#> (Intercept):86 -3.94654029 0.34903165 -11.3071129 0.000000e+00
#> (Intercept):87 -0.85198560 0.28231698  -3.0178333 2.545889e-03
#> (Intercept):88 -1.39318951 0.28527660  -4.8836446 1.041427e-06
#> (Intercept):89 -4.31127179 0.27055041 -15.9351886 0.000000e+00
#> (Intercept):90 -2.56869573 0.35934204  -7.1483307 8.784085e-13
#> (Intercept):91 -5.17367462 0.45117164 -11.4671981 0.000000e+00
#> (Intercept):92 -3.51674997 0.72743938  -4.8344234 1.335321e-06
#> yearsold_1      0.25015794 0.03124184   8.0071439 1.110223e-15
#> yearsold_2     -0.06588471 0.03846166  -1.7129969 8.671310e-02
#> yearsold_3     -0.31528550 0.04879928  -6.4608643 1.041065e-10
#> yearsold_4     -0.64440568 0.06094022 -10.5743908 0.000000e+00
#> yearsold_5     -0.81324927 0.07443488 -10.9256480 0.000000e+00
#> yearsold_6     -0.90376070 0.08642160 -10.4575792 0.000000e+00
#> yearsold_7     -1.00078335 0.10061791  -9.9463739 0.000000e+00
#> yearsold_8     -0.84845898 0.12945244  -6.5542139 5.593570e-11