Duration Data

Duration Data

duration_data_da

Format

dataframe with one row for each rendition of each song in the Fugazi Live Series data.

gid

Unique identifier for the show

date

The date of the show.

song_number

this is the number of the song in the set, where 1 is the first song in that show. Larger numbers will indicate that the song was played later in the set,

song

the name of the song

urls

A string used to form the URLs of the corresponding page on the Fugazi Live series site.

fls_link

a link to the corresponding page of the Fugazi Live Series site.

minutes

duration of the song in minutes

Source

https://www.dischord.com/fugazi_live_series

Examples

  duration_data_da
#> # A tibble: 17,365 × 7
#>    gid                 date       song_number song         urls  fls_l…¹ minutes
#>    <chr>               <date>           <dbl> <chr>        <chr> <chr>     <dbl>
#>  1 aalst-belgium-92390 1990-09-23           2 turnover     http… <a hre…    4.43
#>  2 aalst-belgium-92390 1990-09-23           3 brendan #1   http… <a hre…    2.93
#>  3 aalst-belgium-92390 1990-09-23           4 merchandise  http… <a hre…    3.07
#>  4 aalst-belgium-92390 1990-09-23           5 sieve-fiste… http… <a hre…    3.58
#>  5 aalst-belgium-92390 1990-09-23           6 and the same http… <a hre…    4.3 
#>  6 aalst-belgium-92390 1990-09-23           8 bulldog fro… http… <a hre…    2.75
#>  7 aalst-belgium-92390 1990-09-23           9 burning too  http… <a hre…    2.72
#>  8 aalst-belgium-92390 1990-09-23          11 suggestion   http… <a hre…    6.4 
#>  9 aalst-belgium-92390 1990-09-23          13 reclamation  http… <a hre…    3.52
#> 10 aalst-belgium-92390 1990-09-23          14 blueprint    http… <a hre…    4.33
#> # … with 17,355 more rows, and abbreviated variable name ¹​fls_link