Fugazi song performance intensity data

Fugazi song performance intensity data

fugazi_song_performance_intensity

Format

dataframe with one row for each song in the Fugazi discography, except those which never appear in the Fugazi Live Series data.

songid

numeric id for each song

song

The name of the song

launchdate

The date on which the song was first performed according to the data

chosen

The number of times the song was performed according to the data

available_rl

The number of shows for which the song was available in the band's repertoire

intensity

The performance intensity is the ratio of chosen/available_rl

Source

https://www.dischord.com/fugazi_live_series

Examples

  fugazi_song_performance_intensity
#> # A tibble: 92 x 6
#>    songid song                 launchdate chosen available_rl intensity
#>     <dbl> <chr>                <date>      <dbl>        <dbl>     <dbl>
#>  1     71 cashout              2000-09-30     59           65     0.908
#>  2     44 closed captioned     1997-06-18    158          198     0.798
#>  3     18 bed for the scraping 1994-11-20    299          376     0.795
#>  4     64 number 5             1998-11-21    109          144     0.757
#>  5      2 reclamation          1990-05-05    594          793     0.749
#>  6     42 break                1996-08-15    170          228     0.746
#>  7     46 arpeggiator          1997-05-02    154          207     0.744
#>  8     24 do you like me       1994-11-20    272          376     0.723
#>  9      3 blueprint            1989-11-25    585          823     0.711
#> 10      1 waiting room         1987-09-03    628          895     0.702
#> # ... with 82 more rows