Fugazi Live Series data - other variables

some of this data was scraped from the Fugazi Live Series website by Carni Klirs for his project "Visualizing the History of Fugazi". The original data on coordinates, cities and tours data came from The D-I-Y Data of Fugazi by Matthew Conlen. Rows with checked==1 were updated by Alex Mitrani, in particular making sure that the coordinates indicated the actual locations of the venues for city-level mapping.

othervariables

Format

dataframe with one row for each show.

gid

show id

flsid

Fugazi Live Series id

venue

Venue

doorprice

Door price

attendance

Attendance

recorded_by

Recorded by

mastered_by

Mastered by

original_source

Original source

x

longitude

y

latitude

city

city

country

country

tour

tour

year

year

checked

checked==1 indicates that the data was checked and updated by Alex Mitrani, in particular making sure that the coordinates indicate as closely as possible the actual locations of the venues.

Source

https://www.dischord.com/fugazi_live_series

Examples

othervariables
#> # A tibble: 1,048 × 16
#>    gid    flsid date       venue doorp…¹ atten…² recor…³ maste…⁴ origi…⁵       x
#>    <chr>  <chr> <date>     <chr> <chr>     <dbl> <chr>   <chr>   <chr>     <dbl>
#>  1 aalst… FLS0… 1990-09-23 Netw… ""         600  Joey P… Warren… Casset…    4.04
#>  2 aberd… FLS0… 1999-05-04 Lemo… "6(pou…    550  Joey P… Jerry … DAT       -2.09
#>  3 adela… FLS0… 1993-11-11 Dom … "15"       550  Joey P… Warren… Casset…  139.  
#>  4 adela… FLS0… 1996-11-12 Adel… ""         913. Nick P… Jerry … DAT      139.  
#>  5 adela… FLS0… 1991-10-22 Le R… ""         450  Joey P… Warren… Casset…  139.  
#>  6 akron… FLS0… 1990-06-28 Jack… "5"        700  Joey P… Warren… Casset…  -81.6 
#>  7 alban… FLS0… 1993-09-20 S.U.… "5"       1000  Joey P… Warren… DAT      -73.8 
#>  8 albuq… FLS0… 1995-11-13 Five… "5"        895  Joey P… Warren… DAT     -107.  
#>  9 albuq… FLS0… 2001-04-08 Suns… "6"       1100  Nick P… Warren… DAT     -107.  
#> 10 albuq… FLS0… 1991-09-11 Suns… "5"        800  Joey P… Warren… Casset… -107.  
#> # … with 1,038 more rows, 6 more variables: y <dbl>, city <chr>, country <chr>,
#> #   tour <chr>, year <dbl>, checked <dbl>, and abbreviated variable names
#> #   ¹​doorprice, ²​attendance, ³​recorded_by, ⁴​mastered_by, ⁵​original_source