Undertakes paired comparisons for ranking a set of coefficients, considering whether the differences between the coefficients are significant or not.

The index numbers are based on the model coefficient table that comes straight out of the model, with no sorting.

The function will return a dataframe with the results for each pair of coeeficients tested.

rankr(coeftable = NULL, vcovmat = NULL, mysongidlist = NULL)

Arguments

coeftable

coefficients table from mlogit, with one row per coefficient

vcovmat

variance covariance matrix from mlogit, with one row and one column per coefficient

mysongidlist

a dataframe containing the list of song ids to be tested. It can contain other variables but only songid will be used.

Details

rankr

Examples

songstobecompared <- summary %>% slice(seq(from=1, to=92, by=10))
mycomparisons <- rankr(coeftable = results_ml_Repeatr4, vcovmat = vcovmat_ml_Repeatr4, mysongidlist = songstobecompared)
#> 
#>  
#> First coefficient: 3.61570510729558 
#>  
#> Second coefficient: 3.02276303525277 
#>  
#> Difference to be tested: 0.592942072042805 
#>  
#> Variance of the first coefficient: 0.0046368645007149 
#>  
#> Variance of the second coefficient: 0.0249044783491863 
#>  
#> Covariance of the two coefficients: 0.00336092388537719 
#>  
#> Z-statistic: 3.92517705390937 
#>  
#> P-statistic: 1.99991333403047 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 0.296862062085952 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.889022081999659 
#>  
#> 
#>  
#> First coefficient: 3.02276303525277 
#>  
#> Second coefficient: 2.70579304493306 
#>  
#> Difference to be tested: 0.316969990319709 
#>  
#> Variance of the first coefficient: 0.0249044783491863 
#>  
#> Variance of the second coefficient: 0.00918859680673319 
#>  
#> Covariance of the two coefficients: 0.00196404984620897 
#>  
#> Z-statistic: 1.82501592477836 
#>  
#> P-statistic: 1.93200137375634 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.0234441247527822 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.6573841053922 
#>  
#> 
#>  
#> First coefficient: 2.70579304493306 
#>  
#> Second coefficient: 2.54701258151574 
#>  
#> Difference to be tested: 0.158780463417321 
#>  
#> Variance of the first coefficient: 0.00918859680673319 
#>  
#> Variance of the second coefficient: 0.00443420102574495 
#>  
#> Covariance of the two coefficients: 0.00175090160270639 
#>  
#> Z-statistic: 1.57828514167476 
#>  
#> P-statistic: 1.88549988218511 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.0384017187374933 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.355962645572135 
#>  
#> 
#>  
#> First coefficient: 2.54701258151574 
#>  
#> Second coefficient: 2.33062934218183 
#>  
#> Difference to be tested: 0.216383239333909 
#>  
#> Variance of the first coefficient: 0.00443420102574495 
#>  
#> Variance of the second coefficient: 0.0433360457412497 
#>  
#> Covariance of the two coefficients: 0.00626520531597871 
#>  
#> Z-statistic: 1.15267447958207 
#>  
#> P-statistic: 1.75095597985889 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.151553378164756 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.584319856832574 
#>  
#> 
#>  
#> First coefficient: 2.33062934218183 
#>  
#> Second coefficient: 2.16211755294913 
#>  
#> Difference to be tested: 0.168511789232699 
#>  
#> Variance of the first coefficient: 0.0433360457412497 
#>  
#> Variance of the second coefficient: 0.00939535563032387 
#>  
#> Covariance of the two coefficients: 0.00295334579233787 
#>  
#> Z-statistic: 0.778740466027299 
#>  
#> P-statistic: 1.56386738636208 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.255612962600233 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.592636541065632 
#>  
#> 
#>  
#> First coefficient: 2.16211755294913 
#>  
#> Second coefficient: 1.82086119672103 
#>  
#> Difference to be tested: 0.341256356228107 
#>  
#> Variance of the first coefficient: 0.00939535563032387 
#>  
#> Variance of the second coefficient: 0.0047270624392076 
#>  
#> Covariance of the two coefficients: 0.00175310115331858 
#>  
#> Z-statistic: 3.3120423476799 
#>  
#> P-statistic: 1.99907382474994 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 0.139307713066696 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.543204999389517 
#>  
#> 
#>  
#> First coefficient: 1.82086119672103 
#>  
#> Second coefficient: 1.50659935780833 
#>  
#> Difference to be tested: 0.314261838912695 
#>  
#> Variance of the first coefficient: 0.0047270624392076 
#>  
#> Variance of the second coefficient: 0.00997840548713384 
#>  
#> Covariance of the two coefficients: 0.00189991281618983 
#>  
#> Z-statistic: 3.00930372782271 
#>  
#> P-statistic: 1.9973815279736 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 0.109578875683013 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.518944802142378 
#>  
#> 
#>  
#> First coefficient: 1.50659935780833 
#>  
#> Second coefficient: 0.868536684349473 
#>  
#> Difference to be tested: 0.638062673458857 
#>  
#> Variance of the first coefficient: 0.00997840548713384 
#>  
#> Variance of the second coefficient: 0.0144238577046019 
#>  
#> Covariance of the two coefficients: 0.00520020370201088 
#>  
#> Z-statistic: 5.39225645572618 
#>  
#> P-statistic: 1.99999993042163 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 0.406136975887238 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.869988371030477 
#>  
#> 
#>  
#> First coefficient: 0.868536684349473 
#>  
#> Second coefficient: -1.15083178495992 
#>  
#> Difference to be tested: 2.01936846930939 
#>  
#> Variance of the first coefficient: 0.0144238577046019 
#>  
#> Variance of the second coefficient: 0.0804502488495918 
#>  
#> Covariance of the two coefficients: 0.0155101332101339 
#>  
#> Z-statistic: 7.99138511804788 
#>  
#> P-statistic: 2 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 1.5240898484721 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 2.51464709014669 
#>  
#> 
#> Joining, by = "songid1"
#> Joining, by = "songid2"
mycomparisons
#>                          song1                        song2   mycoef1
#> 1                        break                     number 5 3.6157051
#> 2                     number 5                long division 3.0227630
#> 3                long division                 and the same 2.7057930
#> 4                 and the same                     repeater 2.5470126
#> 5                     repeater                    nightshop 2.3306293
#> 6                    nightshop                  burning too 2.1621176
#> 7                  burning too last chance for a slow dance 1.8208612
#> 8 last chance for a slow dance                    furniture 1.5065994
#> 9                    furniture           turn off your guns 0.8685367
#>      mycoef2 mycoefdiff       myz      myp   lower95ci upper95ci
#> 1  3.0227630  0.5929421 3.9251771 1.999913  0.29686206 0.8890221
#> 2  2.7057930  0.3169700 1.8250159 1.932001 -0.02344412 0.6573841
#> 3  2.5470126  0.1587805 1.5782851 1.885500 -0.03840172 0.3559626
#> 4  2.3306293  0.2163832 1.1526745 1.750956 -0.15155338 0.5843199
#> 5  2.1621176  0.1685118 0.7787405 1.563867 -0.25561296 0.5926365
#> 6  1.8208612  0.3412564 3.3120423 1.999074  0.13930771 0.5432050
#> 7  1.5065994  0.3142618 3.0093037 1.997382  0.10957888 0.5189448
#> 8  0.8685367  0.6380627 5.3922565 2.000000  0.40613698 0.8699884
#> 9 -1.1508318  2.0193685 7.9913851 2.000000  1.52408985 2.5146471