Undertakes paired comparisons for ranking a set of coefficients, considering whether the differences between the coefficients are significant or not.

The index numbers are based on the model coefficient table that comes straight out of the model, with no sorting.

The function will return a dataframe with the results for each pair of coeeficients tested.

rankr(coeftable = NULL, vcovmat = NULL, mysongidlist = NULL)

Arguments

coeftable

coefficients table from mlogit, with one row per coefficient

vcovmat

variance covariance matrix from mlogit, with one row and one column per coefficient

mysongidlist

a dataframe containing the list of song ids to be tested. It can contain other variables but only songid will be used.

Value

Details

rankr

Examples

songstobecompared <- summary %>% slice(seq(from=1, to=92, by=10))
mycomparisons <- rankr(coeftable = results_ml_Repeatr4, vcovmat = vcovmat_ml_Repeatr4, mysongidlist = songstobecompared)
#> 
#>  
#> First coefficient: 0.468594841078476 
#>  
#> Second coefficient: -0.126778620519959 
#>  
#> Difference to be tested: 0.595373461598435 
#>  
#> Variance of the first coefficient: 0.0135838922797678 
#>  
#> Variance of the second coefficient: 0.030078964192655 
#>  
#> Covariance of the two coefficients: 0.0142163979012863 
#>  
#> Z-statistic: 4.82434831041146 
#>  
#> P-statistic: 1.99999859538231 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 0.353489602961079 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.837257320235792 
#>  
#> 
#>  
#> First coefficient: -0.126778620519959 
#>  
#> Second coefficient: -0.433994969238196 
#>  
#> Difference to be tested: 0.307216348718237 
#>  
#> Variance of the first coefficient: 0.030078964192655 
#>  
#> Variance of the second coefficient: 0.0377338209399677 
#>  
#> Covariance of the two coefficients: 0.021417565953137 
#>  
#> Z-statistic: 1.94387577405163 
#>  
#> P-statistic: 1.94808956669094 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.00254832427457108 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.616981021711045 
#>  
#> 
#>  
#> First coefficient: -0.433994969238196 
#>  
#> Second coefficient: -0.582523517498974 
#>  
#> Difference to be tested: 0.148528548260778 
#>  
#> Variance of the first coefficient: 0.0377338209399677 
#>  
#> Variance of the second coefficient: 0.00449433783438713 
#>  
#> Covariance of the two coefficients: 0.00182796814373514 
#>  
#> Z-statistic: 0.756262567262944 
#>  
#> P-statistic: 1.55050821347808 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.236411771666371 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.533468868187927 
#>  
#> 
#>  
#> First coefficient: -0.582523517498974 
#>  
#> Second coefficient: -0.816521209828832 
#>  
#> Difference to be tested: 0.233997692329858 
#>  
#> Variance of the first coefficient: 0.00449433783438713 
#>  
#> Variance of the second coefficient: 0.0047643511516031 
#>  
#> Covariance of the two coefficients: 0.00169450115347429 
#>  
#> Z-statistic: 3.05424670995448 
#>  
#> P-statistic: 1.99774373549538 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 0.0838344866295049 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.38416089803021 
#>  
#> 
#>  
#> First coefficient: -0.816521209828832 
#>  
#> Second coefficient: -1.00182710178585 
#>  
#> Difference to be tested: 0.185305891957023 
#>  
#> Variance of the first coefficient: 0.0047643511516031 
#>  
#> Variance of the second coefficient: 0.0460483564610367 
#>  
#> Covariance of the two coefficients: 0.0049829995347935 
#>  
#> Z-statistic: 0.91687619039276 
#>  
#> P-statistic: 1.64079247816098 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.210821253715919 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.581433037629964 
#>  
#> 
#>  
#> First coefficient: -1.00182710178585 
#>  
#> Second coefficient: -1.3214663931475 
#>  
#> Difference to be tested: 0.319639291361647 
#>  
#> Variance of the first coefficient: 0.0460483564610367 
#>  
#> Variance of the second coefficient: 0.00993069367278609 
#>  
#> Covariance of the two coefficients: 0.00848043868523349 
#>  
#> Z-statistic: 1.61817950188314 
#>  
#> P-statistic: 1.89437608095972 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.067519865171823 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.706798447895117 
#>  
#> 
#>  
#> First coefficient: -1.3214663931475 
#>  
#> Second coefficient: -1.63468830905232 
#>  
#> Difference to be tested: 0.313221915904818 
#>  
#> Variance of the first coefficient: 0.00993069367278609 
#>  
#> Variance of the second coefficient: 0.0113602628441555 
#>  
#> Covariance of the two coefficients: 0.00405114243612707 
#>  
#> Z-statistic: 2.72741744367959 
#>  
#> P-statistic: 1.99361677716432 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 0.0881317095647122 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.538312122244924 
#>  
#> 
#>  
#> First coefficient: -1.63468830905232 
#>  
#> Second coefficient: -2.28160729395168 
#>  
#> Difference to be tested: 0.646918984899359 
#>  
#> Variance of the first coefficient: 0.0113602628441555 
#>  
#> Variance of the second coefficient: 0.0124891074988492 
#>  
#> Covariance of the two coefficients: 0.00168531386513165 
#>  
#> Z-statistic: 4.52062260767317 
#>  
#> P-statistic: 1.99999383419781 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 0.366435228469316 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.927402741329401 
#>  
#> 
#>  
#> First coefficient: -2.28160729395168 
#>  
#> Second coefficient: -4.21526256869906 
#>  
#> Difference to be tested: 1.93365527474738 
#>  
#> Variance of the first coefficient: 0.0124891074988492 
#>  
#> Variance of the second coefficient: 0.0684052560671182 
#>  
#> Covariance of the two coefficients: 0.00169983054164612 
#>  
#> Z-statistic: 6.94613219843986 
#>  
#> P-statistic: 1.99999999999625 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 1.38803301616783 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 2.47927753332692 
#>  
#> 
#> Joining, by = "songid1"
#> Joining, by = "songid2"
mycomparisons
#>                          song1                        song2    mycoef1
#> 1         bed for the scraping                     number 5  0.4685948
#> 2                     number 5                     argument -0.1267786
#> 3                     argument                 and the same -0.4339950
#> 4                 and the same                     repeater -0.5825235
#> 5                     repeater                    nightshop -0.8165212
#> 6                    nightshop                   cassavetes -1.0018271
#> 7                   cassavetes last chance for a slow dance -1.3214664
#> 8 last chance for a slow dance                    furniture -1.6346883
#> 9                    furniture           turn off your guns -2.2816073
#>      mycoef2 mycoefdiff       myz      myp    lower95ci upper95ci
#> 1 -0.1267786  0.5953735 4.8243483 1.999999  0.353489603 0.8372573
#> 2 -0.4339950  0.3072163 1.9438758 1.948090 -0.002548324 0.6169810
#> 3 -0.5825235  0.1485285 0.7562626 1.550508 -0.236411772 0.5334689
#> 4 -0.8165212  0.2339977 3.0542467 1.997744  0.083834487 0.3841609
#> 5 -1.0018271  0.1853059 0.9168762 1.640792 -0.210821254 0.5814330
#> 6 -1.3214664  0.3196393 1.6181795 1.894376 -0.067519865 0.7067984
#> 7 -1.6346883  0.3132219 2.7274174 1.993617  0.088131710 0.5383121
#> 8 -2.2816073  0.6469190 4.5206226 1.999994  0.366435228 0.9274027
#> 9 -4.2152626  1.9336553 6.9461322 2.000000  1.388033016 2.4792775