Undertakes paired comparisons for ranking a set of coefficients, considering whether the differences between the coefficients are significant or not.

The index numbers are based on the model coefficient table that comes straight out of the model, with no sorting.

The function will return a dataframe with the results for each pair of coeeficients tested.

rankr(coeftable = NULL, vcovmat = NULL, mysongidlist = NULL)

Arguments

coeftable

coefficients table from mlogit, with one row per coefficient

vcovmat

variance covariance matrix from mlogit, with one row and one column per coefficient

mysongidlist

a dataframe containing the list of song ids to be tested. It can contain other variables but only songid will be used.

Details

rankr

Examples

songstobecompared <- summary %>% slice(seq(from=1, to=92, by=10))
mycomparisons <- rankr(coeftable = results_ml_Repeatr4, vcovmat = vcovmat_ml_Repeatr4, mysongidlist = songstobecompared)
#> 
#>  
#> First coefficient: 3.15493527070601 
#>  
#> Second coefficient: 0.556095261474314 
#>  
#> Difference to be tested: 2.5988400092317 
#>  
#> Variance of the first coefficient: 0.528235277021886 
#>  
#> Variance of the second coefficient: 0.0124891074988492 
#>  
#> Covariance of the two coefficients: 0.0016387992505919 
#>  
#> Z-statistic: 3.54496383590617 
#>  
#> P-statistic: 1.9996073330434 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 1.16194906932961 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 4.03573094913379 
#>  
#> 
#>  
#> First coefficient: 0.556095261474314 
#>  
#> Second coefficient: 3.01427783881025 
#>  
#> Difference to be tested: -2.45818257733593 
#>  
#> Variance of the first coefficient: 0.0124891074988492 
#>  
#> Variance of the second coefficient: 0.0046794363459201 
#>  
#> Covariance of the two coefficients: 0.00169180892121045 
#>  
#> Z-statistic: -20.9368777231018 
#>  
#> P-statistic: 2.47133325009158e-97 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -2.6883046573948 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: -2.22806049727706 
#>  
#> 
#>  
#> First coefficient: 3.01427783881025 
#>  
#> Second coefficient: 1.57181709608918 
#>  
#> Difference to be tested: 1.44246074272107 
#>  
#> Variance of the first coefficient: 0.0046794363459201 
#>  
#> Variance of the second coefficient: 0.0494846260902958 
#>  
#> Covariance of the two coefficients: 0.00406706571721051 
#>  
#> Z-statistic: 6.72332090223161 
#>  
#> P-statistic: 1.99999999998224 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 1.02195083447098 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 1.86297065097115 
#>  
#> 
#>  
#> First coefficient: 1.57181709608918 
#>  
#> Second coefficient: 1.78114939789564 
#>  
#> Difference to be tested: -0.209332301806466 
#>  
#> Variance of the first coefficient: 0.0494846260902958 
#>  
#> Variance of the second coefficient: 0.0444326554597425 
#>  
#> Covariance of the two coefficients: 0.0206092237692128 
#>  
#> Z-statistic: -0.9118759166017 
#>  
#> P-statistic: 0.361834040751779 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.659274343339608 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.240609739726675 
#>  
#> 
#>  
#> First coefficient: 1.78114939789564 
#>  
#> Second coefficient: 3.43794996598012 
#>  
#> Difference to be tested: -1.65680056808448 
#>  
#> Variance of the first coefficient: 0.0444326554597425 
#>  
#> Variance of the second coefficient: 0.00569784656156396 
#>  
#> Covariance of the two coefficients: 0.00404870824434851 
#>  
#> Z-statistic: -8.08117662629449 
#>  
#> P-statistic: 6.41448252509519e-16 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -2.05863921901551 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: -1.25496191715344 
#>  
#> 
#>  
#> First coefficient: 3.43794996598012 
#>  
#> Second coefficient: 2.20065856399409 
#>  
#> Difference to be tested: 1.23729140198603 
#>  
#> Variance of the first coefficient: 0.00569784656156396 
#>  
#> Variance of the second coefficient: 0.00389868718892827 
#>  
#> Covariance of the two coefficients: 0.0017020834387324 
#>  
#> Z-statistic: 15.7232983652673 
#>  
#> P-statistic: 2 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 1.08305587890713 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 1.39152692506493 
#>  
#> 
#>  
#> First coefficient: 2.20065856399409 
#>  
#> Second coefficient: 2.53415639468825 
#>  
#> Difference to be tested: -0.333497830694164 
#>  
#> Variance of the first coefficient: 0.00389868718892827 
#>  
#> Variance of the second coefficient: 0.0107625693703899 
#>  
#> Covariance of the two coefficients: 0.00171381562879152 
#>  
#> Z-statistic: -3.14653847372104 
#>  
#> P-statistic: 0.00165215523492726 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.541235874776478 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: -0.125759786611849 
#>  
#> 
#>  
#> First coefficient: 2.53415639468825 
#>  
#> Second coefficient: 2.66310935075828 
#>  
#> Difference to be tested: -0.128952956070025 
#>  
#> Variance of the first coefficient: 0.0107625693703899 
#>  
#> Variance of the second coefficient: 0.057980060154878 
#>  
#> Covariance of the two coefficients: 0.00797617930683449 
#>  
#> Z-statistic: -0.561247418990133 
#>  
#> P-statistic: 0.574628881046927 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.579285172032861 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.321379259892811 
#>  
#> 
#>  
#> First coefficient: 2.66310935075828 
#>  
#> Second coefficient: -1.14881660397937 
#>  
#> Difference to be tested: 3.81192595473765 
#>  
#> Variance of the first coefficient: 0.057980060154878 
#>  
#> Variance of the second coefficient: 0.0804502488495918 
#>  
#> Covariance of the two coefficients: 0.0163387855450476 
#>  
#> Z-statistic: 11.7219171113032 
#>  
#> P-statistic: 2 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 3.17454089217981 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 4.44931101729549 
#>  
#> 
#> Joining, by = "songid1"
#> Joining, by = "songid2"
mycomparisons
#>                 song1               song2   mycoef1    mycoef2 mycoefdiff
#> 1        waiting room         provisional 3.1549353  0.5560953  2.5988400
#> 2         provisional           blueprint 0.5560953  3.0142778 -2.4581826
#> 3           blueprint              stacks 3.0142778  1.5718171  1.4424607
#> 4              stacks returning the screw 1.5718171  1.7811494 -0.2093323
#> 5 returning the screw      do you like me 1.7811494  3.4379500 -1.6568006
#> 6      do you like me        back to base 3.4379500  2.2006586  1.2372914
#> 7        back to base        floating boy 2.2006586  2.5341564 -0.3334978
#> 8        floating boy            the kill 2.5341564  2.6631094 -0.1289530
#> 9            the kill  turn off your guns 2.6631094 -1.1488166  3.8119260
#>           myz          myp  lower95ci  upper95ci
#> 1   3.5449638 1.999607e+00  1.1619491  4.0357309
#> 2 -20.9368777 2.471333e-97 -2.6883047 -2.2280605
#> 3   6.7233209 2.000000e+00  1.0219508  1.8629707
#> 4  -0.9118759 3.618340e-01 -0.6592743  0.2406097
#> 5  -8.0811766 6.414483e-16 -2.0586392 -1.2549619
#> 6  15.7232984 2.000000e+00  1.0830559  1.3915269
#> 7  -3.1465385 1.652155e-03 -0.5412359 -0.1257598
#> 8  -0.5612474 5.746289e-01 -0.5792852  0.3213793
#> 9  11.7219171 2.000000e+00  3.1745409  4.4493110