Undertakes paired comparisons for ranking a set of coefficients, considering whether the differences between the coefficients are significant or not.

The index numbers are based on the model coefficient table that comes straight out of the model, with no sorting.

The function will return a dataframe with the results for each pair of coeeficients tested.

rankr(coeftable = NULL, vcovmat = NULL, mysongidlist = NULL)

Arguments

coeftable

coefficients table from mlogit, with one row per coefficient

vcovmat

variance covariance matrix from mlogit, with one row and one column per coefficient

mysongidlist

a dataframe containing the list of song ids to be tested. It can contain other variables but only songid will be used.

Details

rankr

Examples

songstobecompared <- summary %>% slice(seq(from=1, to=92, by=10))
mycomparisons <- rankr(coeftable = results_ml_Repeatr4, vcovmat = vcovmat_ml_Repeatr4, mysongidlist = songstobecompared)
#> 
#>  
#> First coefficient: 0 
#>  
#> Second coefficient: -0.968527242046141 
#>  
#> Difference to be tested: 0.968527242046141 
#>  
#> Variance of the first coefficient: 0 
#>  
#> Variance of the second coefficient: 0.0053793341146143 
#>  
#> Covariance of the two coefficients: 0 
#>  
#> Z-statistic: 13.2052789888661 
#>  
#> P-statistic: 2 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 0.824773111897813 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 1.11228137219447 
#>  
#> 
#>  
#> First coefficient: -0.968527242046141 
#>  
#> Second coefficient: -0.890941358104481 
#>  
#> Difference to be tested: -0.0775858839416597 
#>  
#> Variance of the first coefficient: 0.0053793341146143 
#>  
#> Variance of the second coefficient: 0.0046368645007149 
#>  
#> Covariance of the two coefficients: 0.00173551164616128 
#>  
#> Z-statistic: -0.959007647211479 
#>  
#> P-statistic: 0.337554892752307 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.236154309300753 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.0809825414174335 
#>  
#> 
#>  
#> First coefficient: -0.890941358104481 
#>  
#> Second coefficient: -0.816521209828832 
#>  
#> Difference to be tested: -0.0744201482756491 
#>  
#> Variance of the first coefficient: 0.0046368645007149 
#>  
#> Variance of the second coefficient: 0.0047643511516031 
#>  
#> Covariance of the two coefficients: 0.00206714101622343 
#>  
#> Z-statistic: -1.02544319770585 
#>  
#> P-statistic: 0.30515411435884 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.21666448803684 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.0678241914855413 
#>  
#> 
#>  
#> First coefficient: -0.816521209828832 
#>  
#> Second coefficient: -4.94324764221028 
#>  
#> Difference to be tested: 4.12672643238145 
#>  
#> Variance of the first coefficient: 0.0047643511516031 
#>  
#> Variance of the second coefficient: 0.170138117554029 
#>  
#> Covariance of the two coefficients: 0.00269307638512748 
#>  
#> Z-statistic: 10.0230506968779 
#>  
#> P-statistic: 2 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 3.31974819271365 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 4.93370467204925 
#>  
#> 
#>  
#> First coefficient: -4.94324764221028 
#>  
#> Second coefficient: -1.25648807790034 
#>  
#> Difference to be tested: -3.68675956430994 
#>  
#> Variance of the first coefficient: 0.170138117554029 
#>  
#> Variance of the second coefficient: 0.0129317874230662 
#>  
#> Covariance of the two coefficients: 0.00432219402684357 
#>  
#> Z-statistic: -8.82754227810424 
#>  
#> P-statistic: 1.0700093194989e-18 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -4.50533947232929 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: -2.86817965629059 
#>  
#> 
#>  
#> First coefficient: -1.25648807790034 
#>  
#> Second coefficient: 0.468594841078476 
#>  
#> Difference to be tested: -1.72508291897882 
#>  
#> Variance of the first coefficient: 0.0129317874230662 
#>  
#> Variance of the second coefficient: 0.0135838922797678 
#>  
#> Covariance of the two coefficients: 0.00746515137037676 
#>  
#> Z-statistic: -16.0270979204588 
#>  
#> P-statistic: 8.26536080336916e-58 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -1.93604828139059 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: -1.51411755656705 
#>  
#> 
#>  
#> First coefficient: 0.468594841078476 
#>  
#> Second coefficient: 0.0724564014695536 
#>  
#> Difference to be tested: 0.396138439608922 
#>  
#> Variance of the first coefficient: 0.0135838922797678 
#>  
#> Variance of the second coefficient: 0.0220877859336658 
#>  
#> Covariance of the two coefficients: 0.011941453758839 
#>  
#> Z-statistic: 3.64848635538326 
#>  
#> P-statistic: 1.99973621008995 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 0.183329272736781 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.608947606481064 
#>  
#> 
#>  
#> First coefficient: 0.0724564014695536 
#>  
#> Second coefficient: -0.126778620519959 
#>  
#> Difference to be tested: 0.199235021989513 
#>  
#> Variance of the first coefficient: 0.0220877859336658 
#>  
#> Variance of the second coefficient: 0.030078964192655 
#>  
#> Covariance of the two coefficients: 0.0170527430958592 
#>  
#> Z-statistic: 1.48248945367245 
#>  
#> P-statistic: 1.86178989228077 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: -0.0641736937568888 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 0.462643737735915 
#>  
#> 
#>  
#> First coefficient: -0.126778620519959 
#>  
#> Second coefficient: -3.41225260882984 
#>  
#> Difference to be tested: 3.28547398830988 
#>  
#> Variance of the first coefficient: 0.030078964192655 
#>  
#> Variance of the second coefficient: 0.528235277021886 
#>  
#> Covariance of the two coefficients: 0.0237082308687787 
#>  
#> Z-statistic: 4.59653980752123 
#>  
#> P-statistic: 1.99999570435065 
#>  
#> Lower boundary of 95% confidence interval of the difference between the two coefficients: 1.88452255816008 
#>  
#> Upper boundary of 95% confidence interval of the difference between the two coefficients: 4.68642541845968 
#>  
#> 
#> Joining, by = "songid1"
#> Joining, by = "songid2"
mycomparisons
#>                  song1                song2    mycoef1    mycoef2  mycoefdiff
#> 1         waiting room            bad mouth  0.0000000 -0.9685272  0.96852724
#> 2            bad mouth             promises -0.9685272 -0.8909414 -0.07758588
#> 3             promises             repeater -0.8909414 -0.8165212 -0.07442015
#> 4             repeater               polish -0.8165212 -4.9432476  4.12672643
#> 5               polish    walken's syndrome -4.9432476 -1.2564881 -3.68675956
#> 6    walken's syndrome bed for the scraping -1.2564881  0.4685948 -1.72508292
#> 7 bed for the scraping       place position  0.4685948  0.0724564  0.39613844
#> 8       place position             number 5  0.0724564 -0.1267786  0.19923502
#> 9             number 5        hello morning -0.1267786 -3.4122526  3.28547399
#>           myz          myp   lower95ci   upper95ci
#> 1  13.2052790 2.000000e+00  0.82477311  1.11228137
#> 2  -0.9590076 3.375549e-01 -0.23615431  0.08098254
#> 3  -1.0254432 3.051541e-01 -0.21666449  0.06782419
#> 4  10.0230507 2.000000e+00  3.31974819  4.93370467
#> 5  -8.8275423 1.070009e-18 -4.50533947 -2.86817966
#> 6 -16.0270979 8.265361e-58 -1.93604828 -1.51411756
#> 7   3.6484864 1.999736e+00  0.18332927  0.60894761
#> 8   1.4824895 1.861790e+00 -0.06417369  0.46264374
#> 9   4.5965398 1.999996e+00  1.88452256  4.68642542