Estimated coefficients and related statistics from the model ml.Repeatr4

Basic choice model

Usage

results_ml_Repeatr4

Format

dataframe with one row for each coefficient in the model.

Estimate

The coefficient value

Std. Error

The standard error of the coefficient

z-value

The z-value of the coefficient

Pr(>|z|)

The P value of the coefficient

Details

This model was estimated with mlogit on all of the data.

The utility formula was as follows:

choice ~ yearsold_1 + yearsold_2 + yearsold_3 + yearsold_4 + yearsold_5 + yearsold_6 + yearsold_7 + yearsold_8 + song2 + ... + song92

Provenance

Derived-modeled. Produced by Repeatr_4 (mlogit fit on Repeatr3).

Examples

results_ml_Repeatr4
#>                   Estimate Std. Error     z-value     Pr(>|z|)
#> (Intercept):2   2.53141417 0.20838710  12.1476531 0.000000e+00
#> (Intercept):3   2.70962458 0.23809007  11.3806705 0.000000e+00
#> (Intercept):4   2.85967839 0.21580460  13.2512392 0.000000e+00
#> (Intercept):5   2.19875828 0.21230735  10.3564868 0.000000e+00
#> (Intercept):6   2.25075748 0.20927085  10.7552364 0.000000e+00
#> (Intercept):7   3.60868862 0.20333710  17.7473206 0.000000e+00
#> (Intercept):8   2.75266308 0.20666856  13.3192153 0.000000e+00
#> (Intercept):9   2.99125046 0.20053767  14.9161528 0.000000e+00
#> (Intercept):10  3.56122317 0.21316194  16.7066562 0.000000e+00
#> (Intercept):11  1.23585742 0.21788558   5.6720477 1.411007e-08
#> (Intercept):12  1.50692872 0.21213699   7.1035642 1.215694e-12
#> (Intercept):13  1.94177674 0.20914608   9.2843087 0.000000e+00
#> (Intercept):14  1.87855310 0.21084664   8.9095708 0.000000e+00
#> (Intercept):15  1.88119579 0.20933941   8.9863433 0.000000e+00
#> (Intercept):16  2.54263661 0.20342773  12.4989675 0.000000e+00
#> (Intercept):17  3.24037193 0.24530214  13.2097176 0.000000e+00
#> (Intercept):18  1.83525002 0.20724176   8.8555996 0.000000e+00
#> (Intercept):19  1.54308729 0.24067028   6.4116237 1.439779e-10
#> (Intercept):20  3.31997356 0.21589994  15.3773712 0.000000e+00
#> (Intercept):21  0.07192753 0.29258683   0.2458331 8.058114e-01
#> (Intercept):22  1.59846980 0.21035306   7.5989853 2.975398e-14
#> (Intercept):23  3.43463369 0.20406092  16.8314134 0.000000e+00
#> (Intercept):24  1.91298660 0.21590132   8.8604672 0.000000e+00
#> (Intercept):25  2.25541902 0.27352712   8.2456870 2.220446e-16
#> (Intercept):26  2.40037500 0.25072619   9.5736906 0.000000e+00
#> (Intercept):27  2.17373592 0.20383267  10.6643155 0.000000e+00
#> (Intercept):28  2.76691008 0.20183982  13.7084453 0.000000e+00
#> (Intercept):29  2.60550767 0.22042156  11.8205662 0.000000e+00
#> (Intercept):30  1.82805804 0.21120887   8.6552143 0.000000e+00
#> (Intercept):31  2.52138368 0.22028566  11.4459728 0.000000e+00
#> (Intercept):32  2.45670421 0.22159257  11.0865822 0.000000e+00
#> (Intercept):33  1.20932253 0.26668843   4.5345895 5.771563e-06
#> (Intercept):34  2.56703127 0.20661157  12.4244316 0.000000e+00
#> (Intercept):35  2.45520499 0.28485141   8.6192480 0.000000e+00
#> (Intercept):36  0.94596427 0.22402013   4.2226754 2.414194e-05
#> (Intercept):37  2.40549652 0.20650614  11.6485470 0.000000e+00
#> (Intercept):38  1.14179434 0.21626129   5.2796981 1.293969e-07
#> (Intercept):39  2.40125655 0.20219456  11.8759698 0.000000e+00
#> (Intercept):40  1.74416965 0.21011128   8.3011711 0.000000e+00
#> (Intercept):41  1.07070568 0.26622416   4.0218200 5.775017e-05
#> (Intercept):42 -0.40598183 0.73941681  -0.5490568 5.829665e-01
#> (Intercept):43 -0.40620964 0.27049265  -1.5017400 1.331643e-01
#> (Intercept):44  2.49374795 0.20208262  12.3402397 0.000000e+00
#> (Intercept):45  1.13659629 0.21984655   5.1699528 2.341532e-07
#> (Intercept):46  1.38733121 0.21498473   6.4531617 1.095404e-10
#> (Intercept):47  1.47278294 0.21118033   6.9740537 3.079315e-12
#> (Intercept):48  0.73390729 0.25226942   2.9092202 3.623315e-03
#> (Intercept):49  1.76031236 0.20780820   8.4708514 0.000000e+00
#> (Intercept):50  2.49272683 0.29815551   8.3604923 0.000000e+00
#> (Intercept):51  1.52223635 0.21408684   7.1103687 1.157296e-12
#> (Intercept):52  1.81273029 0.21625051   8.3825481 0.000000e+00
#> (Intercept):53  2.71513040 0.20255900  13.4041460 0.000000e+00
#> (Intercept):54  2.81570081 0.20434211  13.7793466 0.000000e+00
#> (Intercept):55  2.89530578 0.20716765  13.9756657 0.000000e+00
#> (Intercept):56  1.31047716 0.21462793   6.1058090 1.022813e-09
#> (Intercept):57  2.06925487 0.25565525   8.0939269 6.661338e-16
#> (Intercept):58  2.27980022 0.22362611  10.1946959 0.000000e+00
#> (Intercept):59  3.01907229 0.22579312  13.3709669 0.000000e+00
#> (Intercept):60  2.74084983 0.23187539  11.8203565 0.000000e+00
#> (Intercept):61  1.76754624 0.23152675   7.6343068 2.264855e-14
#> (Intercept):62  3.19785928 0.21655571  14.7669130 0.000000e+00
#> (Intercept):63 -1.83762086 0.45199219  -4.0656032 4.790839e-05
#> (Intercept):64  2.20693064 0.20493487  10.7689369 0.000000e+00
#> (Intercept):65  1.18047718 0.37552425   3.1435445 1.669150e-03
#> (Intercept):66  2.69613300 0.20197484  13.3488556 0.000000e+00
#> (Intercept):67  3.11613742 0.21684855  14.3701092 0.000000e+00
#> (Intercept):68  3.60412379 0.19982307  18.0365751 0.000000e+00
#> (Intercept):69  2.89891512 0.20030708  14.4723545 0.000000e+00
#> (Intercept):70  2.33943393 0.20258985  11.5476366 0.000000e+00
#> (Intercept):71  1.86454931 0.20673953   9.0188331 0.000000e+00
#> (Intercept):72  1.77094652 0.21053640   8.4115932 0.000000e+00
#> (Intercept):73  2.01730520 0.20498082   9.8414339 0.000000e+00
#> (Intercept):74  2.18436969 0.20500177  10.6553701 0.000000e+00
#> (Intercept):75  2.95314154 0.20350559  14.5113535 0.000000e+00
#> (Intercept):76  2.58093193 0.20217888  12.7655862 0.000000e+00
#> (Intercept):77  2.43590849 0.20893918  11.6584575 0.000000e+00
#> (Intercept):78  1.55497314 0.20992164   7.4073981 1.287859e-13
#> (Intercept):79  0.30553019 0.24353855   1.2545455 2.096438e-01
#> (Intercept):80  1.70529589 0.31471722   5.4185020 6.010048e-08
#> (Intercept):81  2.16629615 0.20516487  10.5588062 0.000000e+00
#> (Intercept):82  2.31575980 0.20741906  11.1646430 0.000000e+00
#> (Intercept):83  2.12174588 0.20268357  10.4682680 0.000000e+00
#> (Intercept):84  3.13753660 0.20376654  15.3977028 0.000000e+00
#> (Intercept):85  2.61508600 0.27120848   9.6423459 0.000000e+00
#> (Intercept):86 -0.30788459 0.26371385  -1.1674949 2.430105e-01
#> (Intercept):87 -1.02160744 0.31578848  -3.2351004 1.215999e-03
#> (Intercept):88  2.75759231 0.20365455  13.5405387 0.000000e+00
#> (Intercept):89  2.40181315 0.20386387  11.7814558 0.000000e+00
#> (Intercept):90  0.55787539 0.25594221   2.1796928 2.928024e-02
#> (Intercept):91  3.15160718 0.20596079  15.3019763 0.000000e+00
#> (Intercept):92  1.88873395 0.20948519   9.0160740 0.000000e+00
#> yearsold_1      0.24940035 0.02935668   8.4955225 0.000000e+00
#> yearsold_2     -0.05401097 0.03632953  -1.4866960 1.370951e-01
#> yearsold_3     -0.31307437 0.04624427  -6.7700141 1.287703e-11
#> yearsold_4     -0.60905337 0.05737694 -10.6149507 0.000000e+00
#> yearsold_5     -0.77951688 0.07036819 -11.0776886 0.000000e+00
#> yearsold_6     -0.87704880 0.08124530 -10.7950712 0.000000e+00
#> yearsold_7     -0.96521201 0.09492432 -10.1682270 0.000000e+00
#> yearsold_8     -0.82643720 0.12263073  -6.7392343 1.592237e-11