Estimated coefficients and related statistics from the model ml.Repeatr4

Basic choice model

Usage

results_ml_Repeatr4

Format

dataframe with one row for each coefficient in the model.

Estimate

The coefficient value

Std. Error

The standard error of the coefficient

z-value

The z-value of the coefficient

Pr(>|z|)

The P value of the coefficient

Details

This model was estimated with mlogit on all of the data.

The utility formula was as follows:

choice ~ yearsold_1 + yearsold_2 + yearsold_3 + yearsold_4 + yearsold_5 + yearsold_6 + yearsold_7 + yearsold_8 + song2 + ... + song92

Examples

  results_ml_Repeatr4
#>                  Estimate Std. Error     z-value     Pr(>|z|)
#> (Intercept):2   2.5449317 0.21277529  11.9606545 0.000000e+00
#> (Intercept):3   2.7018622 0.24620691  10.9739495 0.000000e+00
#> (Intercept):4   2.8802644 0.22045259  13.0652330 0.000000e+00
#> (Intercept):5   2.2031401 0.21648323  10.1769554 0.000000e+00
#> (Intercept):6   2.1857980 0.21409808  10.2093302 0.000000e+00
#> (Intercept):7   3.6147356 0.20723365  17.4428027 0.000000e+00
#> (Intercept):8   2.7855678 0.21044711  13.2364268 0.000000e+00
#> (Intercept):9   3.0111120 0.20404962  14.7567639 0.000000e+00
#> (Intercept):10  3.6273290 0.21750746  16.6768024 0.000000e+00
#> (Intercept):11  1.2127518 0.22310921   5.4356867 5.458581e-08
#> (Intercept):12  1.4637611 0.21687418   6.7493563 1.485034e-11
#> (Intercept):13  1.8986618 0.21392796   8.8752391 0.000000e+00
#> (Intercept):14  1.7932812 0.21608591   8.2989270 0.000000e+00
#> (Intercept):15  1.8218152 0.21418999   8.5056042 0.000000e+00
#> (Intercept):16  2.5796483 0.20705132  12.4589798 0.000000e+00
#> (Intercept):17  3.3619662 0.25305258  13.2856430 0.000000e+00
#> (Intercept):18  1.8190524 0.21138041   8.6055868 0.000000e+00
#> (Intercept):19  1.5774489 0.24538145   6.4285581 1.288201e-10
#> (Intercept):20  3.3366269 0.22060802  15.1246854 0.000000e+00
#> (Intercept):21  0.1159590 0.29510416   0.3929426 6.943619e-01
#> (Intercept):22  1.6068838 0.21421748   7.5011799 6.328271e-14
#> (Intercept):23  3.4400007 0.20797933  16.5401090 0.000000e+00
#> (Intercept):24  1.9071603 0.22018705   8.6615464 0.000000e+00
#> (Intercept):25  2.3229954 0.28077813   8.2734201 2.220446e-16
#> (Intercept):26  2.3714875 0.26034292   9.1090914 0.000000e+00
#> (Intercept):27  2.1709371 0.20767452  10.4535553 0.000000e+00
#> (Intercept):28  2.7800682 0.20559773  13.5218817 0.000000e+00
#> (Intercept):29  2.6195547 0.22526842  11.6285928 0.000000e+00
#> (Intercept):30  1.8908266 0.21433220   8.8219434 0.000000e+00
#> (Intercept):31  2.5628542 0.22464145  11.4086433 0.000000e+00
#> (Intercept):32  2.5381197 0.22541310  11.2598588 0.000000e+00
#> (Intercept):33  1.2672564 0.27165818   4.6648930 3.087777e-06
#> (Intercept):34  2.6094116 0.21027066  12.4097753 0.000000e+00
#> (Intercept):35  2.4223096 0.29540541   8.1999503 2.220446e-16
#> (Intercept):36  0.8610537 0.23080985   3.7305761 1.910424e-04
#> (Intercept):37  2.3792718 0.21101869  11.2751707 0.000000e+00
#> (Intercept):38  1.0905833 0.22171483   4.9188561 8.705145e-07
#> (Intercept):39  2.4096372 0.20601075  11.6966573 0.000000e+00
#> (Intercept):40  1.7203389 0.21453794   8.0188097 1.110223e-15
#> (Intercept):41  1.1528479 0.26923216   4.2819844 1.852340e-05
#> (Intercept):42 -0.2840934 0.74088949  -0.3834491 7.013868e-01
#> (Intercept):43 -0.5671355 0.28744951  -1.9729918 4.849650e-02
#> (Intercept):44  2.5003213 0.20592385  12.1419700 0.000000e+00
#> (Intercept):45  1.1760132 0.22421428   5.2450415 1.562472e-07
#> (Intercept):46  1.3989762 0.21963153   6.3696510 1.894589e-10
#> (Intercept):47  1.4983507 0.21481144   6.9751905 3.054668e-12
#> (Intercept):48  0.7809690 0.25518419   3.0604130 2.210320e-03
#> (Intercept):49  1.7588546 0.21169526   8.3084271 0.000000e+00
#> (Intercept):50  2.4238423 0.31144394   7.7825958 7.105427e-15
#> (Intercept):51  1.4688784 0.21943161   6.6940149 2.171285e-11
#> (Intercept):52  1.8505193 0.21982199   8.4182628 0.000000e+00
#> (Intercept):53  2.6994533 0.20666364  13.0620622 0.000000e+00
#> (Intercept):54  2.7782867 0.20867823  13.3137352 0.000000e+00
#> (Intercept):55  2.9489128 0.21143910  13.9468658 0.000000e+00
#> (Intercept):56  1.3077889 0.21867976   5.9803837 2.226126e-09
#> (Intercept):57  2.1632358 0.26130640   8.2785413 2.220446e-16
#> (Intercept):58  2.2696531 0.22921529   9.9018401 0.000000e+00
#> (Intercept):59  3.0267289 0.23144790  13.0773660 0.000000e+00
#> (Intercept):60  2.6848145 0.23902575  11.2323233 0.000000e+00
#> (Intercept):61  1.8093470 0.23578817   7.6736122 1.665335e-14
#> (Intercept):62  3.2311148 0.22113987  14.6111819 0.000000e+00
#> (Intercept):63 -1.8026276 0.45360442  -3.9740081 7.067317e-05
#> (Intercept):64 -2.0675196 0.45567307  -4.5372872 5.698247e-06
#> (Intercept):65  2.2472968 0.20905518  10.7497779 0.000000e+00
#> (Intercept):66  0.5217540 0.41017164   1.2720383 2.033595e-01
#> (Intercept):67  2.7002325 0.20578091  13.1218802 0.000000e+00
#> (Intercept):68  3.1354875 0.22176582  14.1387319 0.000000e+00
#> (Intercept):69  3.5902420 0.20368454  17.6264828 0.000000e+00
#> (Intercept):70  2.9297788 0.20402713  14.3597513 0.000000e+00
#> (Intercept):71  2.3284697 0.20660428  11.2701912 0.000000e+00
#> (Intercept):72  1.8784109 0.21079662   8.9110107 0.000000e+00
#> (Intercept):73  1.7811289 0.21440197   8.3074278 0.000000e+00
#> (Intercept):74  2.0103313 0.20891235   9.6228459 0.000000e+00
#> (Intercept):75  2.1654502 0.20924249  10.3489984 0.000000e+00
#> (Intercept):76  2.9622055 0.20758545  14.2698126 0.000000e+00
#> (Intercept):77  2.5721154 0.20611149  12.4792433 0.000000e+00
#> (Intercept):78  2.3903059 0.21361782  11.1896372 0.000000e+00
#> (Intercept):79  1.5696995 0.21367022   7.3463655 2.036149e-13
#> (Intercept):80  0.3453195 0.24650714   1.4008498 1.612590e-01
#> (Intercept):81  1.8285532 0.31812219   5.7479585 9.032739e-09
#> (Intercept):82  2.2740648 0.20771876  10.9478062 0.000000e+00
#> (Intercept):83  2.2700402 0.21204313  10.7055585 0.000000e+00
#> (Intercept):84  2.1415705 0.20637145  10.3772617 0.000000e+00
#> (Intercept):85  3.1535223 0.20760120  15.1902897 0.000000e+00
#> (Intercept):86  2.6679460 0.27792366   9.5995642 0.000000e+00
#> (Intercept):87 -0.4397045 0.27780892  -1.5827587 1.134765e-01
#> (Intercept):88 -1.0880002 0.33057459  -3.2912397 9.974686e-04
#> (Intercept):89  2.7509584 0.20779602  13.2387443 0.000000e+00
#> (Intercept):90  2.3799839 0.20803209  11.4404653 0.000000e+00
#> (Intercept):91  0.5465618 0.26194008   2.0865911 3.692512e-02
#> (Intercept):92  3.1479071 0.21032042  14.9671965 0.000000e+00
#> (Intercept):93  1.8860891 0.21354074   8.8324554 0.000000e+00
#> (Intercept):94 -1.7970604 0.73523845  -2.4441872 1.451789e-02
#> yearsold_1      0.2061576 0.03014010   6.8399784 7.920553e-12
#> yearsold_2     -0.0833528 0.03725101  -2.2375986 2.524724e-02
#> yearsold_3     -0.3260297 0.04739968  -6.8783112 6.056711e-12
#> yearsold_4     -0.6192220 0.05928850 -10.4442177 0.000000e+00
#> yearsold_5     -0.7830718 0.07256875 -10.7907575 0.000000e+00
#> yearsold_6     -0.8704713 0.08412720 -10.3470854 0.000000e+00
#> yearsold_7     -0.9692983 0.09819715  -9.8709408 0.000000e+00
#> yearsold_8     -0.8092977 0.12638083  -6.4036427 1.517129e-10