Estimated coefficients and related statistics from the model ml.Repeatr4

Basic choice model

results_ml_Repeatr4

Format

dataframe with one row for each coefficient in the model.

Estimate

The coefficient value

Std. Error

The standard error of the coefficient

z-value

The z-value of the coefficient

Pr(>|z|)

The P value of the coefficient

Details

This model was estimated with mlogit on all of the data.

The utility formula was as follows:

choice ~ yearsold_1 + yearsold_2 + yearsold_3 + yearsold_4 + yearsold_5 + yearsold_6 + yearsold_7 + yearsold_8 + song2 + ... + song92

Examples

  results_ml_Repeatr4
#>                   Estimate Std. Error     z-value     Pr(>|z|)
#> (Intercept):2   2.54708854 0.21280546  11.9690939 0.000000e+00
#> (Intercept):3   2.69769171 0.24622264  10.9563104 0.000000e+00
#> (Intercept):4   2.87698892 0.22046308  13.0497540 0.000000e+00
#> (Intercept):5   2.20118809 0.21648563  10.1678255 0.000000e+00
#> (Intercept):6   2.18733588 0.21413948  10.2145383 0.000000e+00
#> (Intercept):7   3.61327076 0.20723702  17.4354500 0.000000e+00
#> (Intercept):8   2.78434132 0.21044934  13.2304591 0.000000e+00
#> (Intercept):9   3.01364584 0.20406076  14.7683753 0.000000e+00
#> (Intercept):10  3.61513085 0.21758267  16.6149760 0.000000e+00
#> (Intercept):11  1.21975477 0.22311749   5.4668722 4.580461e-08
#> (Intercept):12  1.47034369 0.21688438   6.7793896 1.206857e-11
#> (Intercept):13  1.89921683 0.21398370   8.8755212 0.000000e+00
#> (Intercept):14  1.80139831 0.21610228   8.3358598 0.000000e+00
#> (Intercept):15  1.82156691 0.21425003   8.5020613 0.000000e+00
#> (Intercept):16  2.57934585 0.20705169  12.4574970 0.000000e+00
#> (Intercept):17  3.32193972 0.25354135  13.1021613 0.000000e+00
#> (Intercept):18  1.82019675 0.21138155   8.6109538 0.000000e+00
#> (Intercept):19  1.57491190 0.24538633   6.4180914 1.379934e-10
#> (Intercept):20  3.33328288 0.22061952  15.1087393 0.000000e+00
#> (Intercept):21  0.11404639 0.29510620   0.3864588 6.991569e-01
#> (Intercept):22  1.60865616 0.21421969   7.5093759 5.950795e-14
#> (Intercept):23  3.43844354 0.20798281  16.5323445 0.000000e+00
#> (Intercept):24  1.90529887 0.22018989   8.6529808 0.000000e+00
#> (Intercept):25  2.31950594 0.28080002   8.2603483 2.220446e-16
#> (Intercept):26  2.36815156 0.26036294   9.0955783 0.000000e+00
#> (Intercept):27  2.17316763 0.20767785  10.4641283 0.000000e+00
#> (Intercept):28  2.78128295 0.20559902  13.5277056 0.000000e+00
#> (Intercept):29  2.61629948 0.22527855  11.6136202 0.000000e+00
#> (Intercept):30  1.89042059 0.21433280   8.8200245 0.000000e+00
#> (Intercept):31  2.55993792 0.22464847  11.3953052 0.000000e+00
#> (Intercept):32  2.53503357 0.22542002  11.2458226 0.000000e+00
#> (Intercept):33  1.26302943 0.27167028   4.6491262 3.333442e-06
#> (Intercept):34  2.60808779 0.21027268  12.4033604 0.000000e+00
#> (Intercept):35  2.36112118 0.29795355   7.9244607 2.220446e-15
#> (Intercept):36  0.86825492 0.23082027   3.7616061 1.688258e-04
#> (Intercept):37  2.38431456 0.21104982  11.2974018 0.000000e+00
#> (Intercept):38  1.09768048 0.22172553   4.9506273 7.397466e-07
#> (Intercept):39  2.41060309 0.20601158  11.7012988 0.000000e+00
#> (Intercept):40  1.72715991 0.21454827   8.0502161 8.881784e-16
#> (Intercept):41  1.14986092 0.26923763   4.2708032 1.947702e-05
#> (Intercept):42 -0.28639095 0.74090551  -0.3865418 6.990955e-01
#> (Intercept):43 -0.56023109 0.28745779  -1.9489160 5.130545e-02
#> (Intercept):44  2.50125097 0.20592472  12.1464337 0.000000e+00
#> (Intercept):45  1.17294468 0.22435773   5.2280109 1.713434e-07
#> (Intercept):46  1.40590576 0.21964095   6.4009275 1.544360e-10
#> (Intercept):47  1.49950638 0.21481255   6.9805344 2.940537e-12
#> (Intercept):48  0.77908640 0.25518652   3.0530077 2.265601e-03
#> (Intercept):49  1.76083699 0.21169796   8.3176852 0.000000e+00
#> (Intercept):50  2.41120479 0.31143440   7.7422559 9.769963e-15
#> (Intercept):51  1.46852220 0.21952533   6.6895343 2.238809e-11
#> (Intercept):52  1.84863267 0.21982470   8.4095765 0.000000e+00
#> (Intercept):53  2.70642738 0.20667410  13.0951452 0.000000e+00
#> (Intercept):54  2.78072331 0.20870068  13.3239780 0.000000e+00
#> (Intercept):55  2.95790343 0.21144835  13.9887751 0.000000e+00
#> (Intercept):56  1.30950427 0.21868176   5.9881733 2.122109e-09
#> (Intercept):57  2.16165210 0.26133224   8.2716626 2.220446e-16
#> (Intercept):58  2.26663530 0.22922199   9.8883853 0.000000e+00
#> (Intercept):59  3.02273452 0.23146755  13.0589993 0.000000e+00
#> (Intercept):60  2.68089371 0.23904716  11.2149156 0.000000e+00
#> (Intercept):61  1.80667245 0.23579339   7.6620996 1.820766e-14
#> (Intercept):62  3.22846563 0.22114847  14.5986343 0.000000e+00
#> (Intercept):63 -1.80092625 0.45360536  -3.9702490 7.179755e-05
#> (Intercept):64 -2.06086059 0.45567469  -4.5226576 6.106798e-06
#> (Intercept):65  2.25456493 0.20905887  10.7843545 0.000000e+00
#> (Intercept):66  0.55512067 0.41025116   1.3531239 1.760160e-01
#> (Intercept):67  2.70014022 0.20578112  13.1214187 0.000000e+00
#> (Intercept):68  3.13228538 0.22177715  14.1235711 0.000000e+00
#> (Intercept):69  3.59287344 0.20368943  17.6389784 0.000000e+00
#> (Intercept):70  2.93072326 0.20402805  14.3643152 0.000000e+00
#> (Intercept):71  2.33115683 0.20662457  11.2820890 0.000000e+00
#> (Intercept):72  1.88053563 0.21083870   8.9193094 0.000000e+00
#> (Intercept):73  1.78113998 0.21440199   8.3074786 0.000000e+00
#> (Intercept):74  2.01300824 0.20891729   9.6354315 0.000000e+00
#> (Intercept):75  2.16714322 0.20927484  10.3554887 0.000000e+00
#> (Intercept):76  2.96387692 0.20760823  14.2762975 0.000000e+00
#> (Intercept):77  2.56743651 0.20614938  12.4542531 0.000000e+00
#> (Intercept):78  2.39518868 0.21364947  11.2108337 0.000000e+00
#> (Intercept):79  1.57142606 0.21367228   7.3543751 1.918465e-13
#> (Intercept):80  0.34692932 0.24650869   1.4073716 1.593172e-01
#> (Intercept):81  1.82420213 0.31814566   5.7338584 9.817132e-09
#> (Intercept):82  2.27647259 0.20772288  10.9591809 0.000000e+00
#> (Intercept):83  2.27228599 0.21207414  10.7145833 0.000000e+00
#> (Intercept):84  2.14212057 0.20637172  10.3799135 0.000000e+00
#> (Intercept):85  3.15226210 0.20760350  15.1840506 0.000000e+00
#> (Intercept):86  2.66480595 0.27795595   9.5871521 0.000000e+00
#> (Intercept):87 -0.43290088 0.27781948  -1.5582092 1.191837e-01
#> (Intercept):88 -1.08112825 0.33058184  -3.2703801 1.074031e-03
#> (Intercept):89  2.75203943 0.20781964  13.2424418 0.000000e+00
#> (Intercept):90  2.38276432 0.20805821  11.4523925 0.000000e+00
#> (Intercept):91  0.54519260 0.26194171   2.0813508 3.740181e-02
#> (Intercept):92  3.15285955 0.21033936  14.9893938 0.000000e+00
#> (Intercept):93  1.88618100 0.21354084   8.8328817 0.000000e+00
#> (Intercept):94 -1.79969250 0.73524002  -2.4477619 1.437466e-02
#> yearsold_1      0.20390148 0.03017607   6.7570590 1.408207e-11
#> yearsold_2     -0.08562067 0.03729582  -2.2957178 2.169202e-02
#> yearsold_3     -0.32907945 0.04742909  -6.9383463 3.967271e-12
#> yearsold_4     -0.62352443 0.05932328 -10.5106192 0.000000e+00
#> yearsold_5     -0.78872375 0.07262561 -10.8601328 0.000000e+00
#> yearsold_6     -0.87628823 0.08417980 -10.4097215 0.000000e+00
#> yearsold_7     -0.97498428 0.09826967  -9.9215174 0.000000e+00
#> yearsold_8     -0.81818640 0.12647061  -6.4693795 9.840617e-11