Estimated coefficients and related statistics from the model ml.Repeatr4

Basic choice model

results_ml_Repeatr4

Format

dataframe with one row for each coefficient in the model.

Estimate

The coefficient value

Std. Error

The standard error of the coefficient

z-value

The z-value of the coefficient

Pr(>|z|)

The P value of the coefficient

Details

This model was estimated with mlogit on all of the data.

The utility formula was as follows:

choice ~ yearsold_1 + yearsold_2 + yearsold_3 + yearsold_4 + yearsold_5 + yearsold_6 + yearsold_7 + yearsold_8 + song2 + ... + song92

Examples

  results_ml_Repeatr4
#>                   Estimate Std. Error     z-value     Pr(>|z|)
#> (Intercept):2   2.54903524 0.21280613  11.9782042 0.000000e+00
#> (Intercept):3   2.69633314 0.24622232  10.9508073 0.000000e+00
#> (Intercept):4   2.87609519 0.22046298  13.0457057 0.000000e+00
#> (Intercept):5   2.20065856 0.21648557  10.1653821 0.000000e+00
#> (Intercept):6   2.18920111 0.21413945  10.2232498 0.000000e+00
#> (Intercept):7   3.61284503 0.20723693  17.4334036 0.000000e+00
#> (Intercept):8   2.78392230 0.21044936  13.2284664 0.000000e+00
#> (Intercept):9   3.01427784 0.20406104  14.7714518 0.000000e+00
#> (Intercept):10  3.61434845 0.21758269  16.6113783 0.000000e+00
#> (Intercept):11  1.22145832 0.22311765   5.4745033 4.387405e-08
#> (Intercept):12  1.47111529 0.21688472   6.7829366 1.177569e-11
#> (Intercept):13  1.90084280 0.21398240   8.8831736 0.000000e+00
#> (Intercept):14  1.80310944 0.21610202   8.3437878 0.000000e+00
#> (Intercept):15  1.82247951 0.21425039   8.5063068 0.000000e+00
#> (Intercept):16  2.57928094 0.20705171  12.4571826 0.000000e+00
#> (Intercept):17  3.32032358 0.25354068  13.0958220 0.000000e+00
#> (Intercept):18  1.82047741 0.21138157   8.6122805 0.000000e+00
#> (Intercept):19  1.57423221 0.24538619   6.4153253 1.405225e-10
#> (Intercept):20  3.33226957 0.22061939  15.1041554 0.000000e+00
#> (Intercept):21  0.11355482 0.29510616   0.3847931 7.003907e-01
#> (Intercept):22  1.60906824 0.21421974   7.5112977 5.861978e-14
#> (Intercept):23  3.43794997 0.20798276  16.5299760 0.000000e+00
#> (Intercept):24  1.90481643 0.22018983   8.6507922 0.000000e+00
#> (Intercept):25  2.31788525 0.28079947   8.2545926 2.220446e-16
#> (Intercept):26  2.36701375 0.26036273   9.0912157 0.000000e+00
#> (Intercept):27  2.17368176 0.20767796  10.4665980 0.000000e+00
#> (Intercept):28  2.78158061 0.20559904  13.5291521 0.000000e+00
#> (Intercept):29  2.61535876 0.22527838  11.6094531 0.000000e+00
#> (Intercept):30  1.89026621 0.21433284   8.8193027 0.000000e+00
#> (Intercept):31  2.55912048 0.22464845  11.3916675 0.000000e+00
#> (Intercept):32  2.53415639 0.22541986  11.2419394 0.000000e+00
#> (Intercept):33  1.26196151 0.27167017   4.6451971 3.397519e-06
#> (Intercept):34  2.60768882 0.21027271  12.4014610 0.000000e+00
#> (Intercept):35  2.35942625 0.29795305   7.9187854 2.442491e-15
#> (Intercept):36  0.87057364 0.23082540   3.7715678 1.622251e-04
#> (Intercept):37  2.38328509 0.21106679  11.2916157 0.000000e+00
#> (Intercept):38  1.09919309 0.22172498   4.9574616 7.142015e-07
#> (Intercept):39  2.41080683 0.20601164  11.7022844 0.000000e+00
#> (Intercept):40  1.72792732 0.21454860   8.0537803 8.881784e-16
#> (Intercept):41  1.14905667 0.26923762   4.2678162 1.973959e-05
#> (Intercept):42 -0.28813200 0.74090516  -0.3888919 6.973561e-01
#> (Intercept):43 -0.55855606 0.28745842  -1.9430847 5.200592e-02
#> (Intercept):44  2.50144537 0.20592479  12.1473738 0.000000e+00
#> (Intercept):45  1.17468896 0.22435830   5.2357723 1.642964e-07
#> (Intercept):46  1.40758048 0.21964156   6.4085344 1.469251e-10
#> (Intercept):47  1.49977924 0.21481259   6.9818034 2.914113e-12
#> (Intercept):48  0.77859365 0.25518646   3.0510774 2.280218e-03
#> (Intercept):49  1.76130029 0.21169803   8.3198707 0.000000e+00
#> (Intercept):50  2.40946772 0.31143367   7.7366963 1.021405e-14
#> (Intercept):51  1.47022737 0.21952521   6.6973055 2.122968e-11
#> (Intercept):52  1.84814268 0.21982464   8.4073499 0.000000e+00
#> (Intercept):53  2.70720299 0.20667444  13.0988765 0.000000e+00
#> (Intercept):54  2.78164600 0.20870107  13.3283741 0.000000e+00
#> (Intercept):55  2.96002317 0.21144878  13.9987715 0.000000e+00
#> (Intercept):56  1.30989636 0.21868180   5.9899650 2.098862e-09
#> (Intercept):57  2.16045424 0.26133208   8.2670840 2.220446e-16
#> (Intercept):58  2.26583439 0.22922202   9.8848896 0.000000e+00
#> (Intercept):59  3.02143755 0.23146738  13.0534056 0.000000e+00
#> (Intercept):60  2.67970616 0.23904683  11.2099634 0.000000e+00
#> (Intercept):61  1.80597386 0.23579330   7.6591399 1.865175e-14
#> (Intercept):62  3.22759930 0.22114833  14.5947260 0.000000e+00
#> (Intercept):63 -1.80054466 0.45360539  -3.9694076 7.205155e-05
#> (Intercept):64 -2.06002541 0.45567484  -4.5208232 6.159961e-06
#> (Intercept):65  2.25542427 0.20905921  10.7884473 0.000000e+00
#> (Intercept):66  0.55609526 0.41025122   1.3554993 1.752585e-01
#> (Intercept):67  2.70010069 0.20578113  13.1212258 0.000000e+00
#> (Intercept):68  3.13134625 0.22177694  14.1193499 0.000000e+00
#> (Intercept):69  3.59348892 0.20368965  17.6419813 0.000000e+00
#> (Intercept):70  2.93090701 0.20402811  14.3652121 0.000000e+00
#> (Intercept):71  2.33184771 0.20662487  11.2854165 0.000000e+00
#> (Intercept):72  1.88133085 0.21083906   8.9230660 0.000000e+00
#> (Intercept):73  1.78114940 0.21440200   8.3075222 0.000000e+00
#> (Intercept):74  2.01363743 0.20891754   9.6384319 0.000000e+00
#> (Intercept):75  2.16791921 0.20927517  10.3591801 0.000000e+00
#> (Intercept):76  2.96467429 0.20760858  14.2801144 0.000000e+00
#> (Intercept):77  2.56744987 0.20614939  12.4543170 0.000000e+00
#> (Intercept):78  2.39713267 0.21364985  11.2199129 0.000000e+00
#> (Intercept):79  1.57181710 0.21367233   7.3562034 1.891820e-13
#> (Intercept):80  0.34729541 0.24650872   1.4088565 1.588776e-01
#> (Intercept):81  1.82248772 0.31814494   5.7284825 1.013330e-08
#> (Intercept):82  2.27703937 0.20772306  10.9618997 0.000000e+00
#> (Intercept):83  2.27401325 0.21207381  10.7227445 0.000000e+00
#> (Intercept):84  2.14223806 0.20637172  10.3804824 0.000000e+00
#> (Intercept):85  3.15189045 0.20760356  15.1822561 0.000000e+00
#> (Intercept):86  2.66310935 0.27795511   9.5810770 0.000000e+00
#> (Intercept):87 -0.43122461 0.27782008  -1.5521722 1.206210e-01
#> (Intercept):88 -1.14881660 0.33770778  -3.4018068 6.694195e-04
#> (Intercept):89  2.74895769 0.20783455  13.2266637 0.000000e+00
#> (Intercept):90  2.38353179 0.20805853  11.4560637 0.000000e+00
#> (Intercept):91  0.54477940 0.26194173   2.0797733 3.754633e-02
#> (Intercept):92  3.15493527 0.21034001  14.9992163 0.000000e+00
#> (Intercept):93  1.88619041 0.21354085   8.8329254 0.000000e+00
#> (Intercept):94 -1.80038108 0.73523997  -2.4486986 1.433734e-02
#> yearsold_1      0.20369606 0.03017622   6.7502190 1.476219e-11
#> yearsold_2     -0.08606173 0.03729580  -2.3075445 2.102448e-02
#> yearsold_3     -0.32978097 0.04742972  -6.9530452 3.574918e-12
#> yearsold_4     -0.62434223 0.05932352 -10.5243630 0.000000e+00
#> yearsold_5     -0.78992166 0.07262649 -10.8764945 0.000000e+00
#> yearsold_6     -0.87744126 0.08417983 -10.4234145 0.000000e+00
#> yearsold_7     -0.97661152 0.09827031  -9.9380117 0.000000e+00
#> yearsold_8     -0.82049910 0.12647715  -6.4873308 8.737033e-11