Estimated coefficients and related statistics from the model ml.Repeatr4

Basic choice model

results_ml_Repeatr4

Format

dataframe with one row for each coefficient in the model.

Estimate

The coefficient value

Std. Error

The standard error of the coefficient

z-value

The z-value of the coefficient

Pr(>|z|)

The P value of the coefficient

Details

This model was estimated with mlogit on all of the data.

The utility formula was as follows:

choice ~ yearsold_1 + yearsold_2 + yearsold_3 + yearsold_4 + yearsold_5 + yearsold_6 + yearsold_7 + yearsold_8 + song2 + ... + song92

Examples

  results_ml_Repeatr4
#>                   Estimate Std. Error     z-value     Pr(>|z|)
#> (Intercept):2   2.54831783 0.21280539  11.9748747 0.000000e+00
#> (Intercept):3   2.69625600 0.24622225  10.9504971 0.000000e+00
#> (Intercept):4   2.87604617 0.22046296  13.0454844 0.000000e+00
#> (Intercept):5   2.20063000 0.21648556  10.1652509 0.000000e+00
#> (Intercept):6   2.18852242 0.21413961  10.2200729 0.000000e+00
#> (Intercept):7   3.61281924 0.20723692  17.4332803 0.000000e+00
#> (Intercept):8   2.78389788 0.21044935  13.2283510 0.000000e+00
#> (Intercept):9   3.01431446 0.20406101  14.7716339 0.000000e+00
#> (Intercept):10  3.61430635 0.21758269  16.6111851 0.000000e+00
#> (Intercept):11  1.22096478 0.22311758   5.4722930 4.442496e-08
#> (Intercept):12  1.47120466 0.21688462   6.7833518 1.174194e-11
#> (Intercept):13  1.90022553 0.21398393   8.8802255 0.000000e+00
#> (Intercept):14  1.80253944 0.21610248   8.3411324 0.000000e+00
#> (Intercept):15  1.82255618 0.21425028   8.5066688 0.000000e+00
#> (Intercept):16  2.57927348 0.20705171  12.4571466 0.000000e+00
#> (Intercept):17  3.32022931 0.25354056  13.0954564 0.000000e+00
#> (Intercept):18  1.82049412 0.21138157   8.6123599 0.000000e+00
#> (Intercept):19  1.57420474 0.24538618   6.4152136 1.406255e-10
#> (Intercept):20  3.33221716 0.22061938  15.1039183 0.000000e+00
#> (Intercept):21  0.11352568 0.29510615   0.3846944 7.004639e-01
#> (Intercept):22  1.60909349 0.21421973   7.5114160 5.839773e-14
#> (Intercept):23  3.43791821 0.20798274  16.5298244 0.000000e+00
#> (Intercept):24  1.90478629 0.22018981   8.6506559 0.000000e+00
#> (Intercept):25  2.31779126 0.28079936   8.2542613 2.220446e-16
#> (Intercept):26  2.36693376 0.26036268   9.0909103 0.000000e+00
#> (Intercept):27  2.17372750 0.20767791  10.4668208 0.000000e+00
#> (Intercept):28  2.78159735 0.20559903  13.5292340 0.000000e+00
#> (Intercept):29  2.61531187 0.22527836  11.6092461 0.000000e+00
#> (Intercept):30  1.89025559 0.21433284   8.8192531 0.000000e+00
#> (Intercept):31  2.55908370 0.22464844  11.3915044 0.000000e+00
#> (Intercept):32  2.53411620 0.22541984  11.2417622 0.000000e+00
#> (Intercept):33  1.26190049 0.27167014   4.6449730 3.401209e-06
#> (Intercept):34  2.60766455 0.21027270  12.4013463 0.000000e+00
#> (Intercept):35  2.35932712 0.29795290   7.9184566 2.442491e-15
#> (Intercept):36  0.88089106 0.23058226   3.8202899 1.332949e-04
#> (Intercept):37  2.38542736 0.21105006  11.3026611 0.000000e+00
#> (Intercept):38  1.09876780 0.22172569   4.9555277 7.213426e-07
#> (Intercept):39  2.41081969 0.20601164  11.7023471 0.000000e+00
#> (Intercept):40  1.72801116 0.21454851   8.0541744 8.881784e-16
#> (Intercept):41  1.14902250 0.26923761   4.2676894 1.975080e-05
#> (Intercept):42 -0.28823030 0.74090509  -0.3890246 6.972580e-01
#> (Intercept):43 -0.55897295 0.28745780  -1.9445391 5.183047e-02
#> (Intercept):44  2.50145686 0.20592478  12.1474299 0.000000e+00
#> (Intercept):45  1.17420129 0.22435776   5.2336113 1.662298e-07
#> (Intercept):46  1.40713576 0.21964099   6.4065261 1.488727e-10
#> (Intercept):47  1.49979762 0.21481258   6.9818891 2.912337e-12
#> (Intercept):48  0.77856475 0.25518645   3.0509643 2.281077e-03
#> (Intercept):49  1.76134262 0.21169799   8.3200725 0.000000e+00
#> (Intercept):50  2.40936725 0.31143352   7.7363774 1.021405e-14
#> (Intercept):51  1.46969401 0.21952546   6.6948683 2.158651e-11
#> (Intercept):52  1.84811369 0.21982463   8.4072186 0.000000e+00
#> (Intercept):53  2.70727755 0.20667435  13.0992433 0.000000e+00
#> (Intercept):54  2.78170084 0.20870097  13.3286434 0.000000e+00
#> (Intercept):55  2.95913084 0.21144832  13.9945820 0.000000e+00
#> (Intercept):56  1.30992287 0.21868179   5.9900866 2.097293e-09
#> (Intercept):57  2.16037472 0.26133200   8.2667822 2.220446e-16
#> (Intercept):58  2.26579702 0.22922201   9.8847272 0.000000e+00
#> (Intercept):59  3.02137010 0.23146732  13.0531172 0.000000e+00
#> (Intercept):60  2.67963768 0.23904676  11.2096799 0.000000e+00
#> (Intercept):61  1.80594235 0.23579328   7.6590067 1.865175e-14
#> (Intercept):62  3.22755511 0.22114833  14.5945262 0.000000e+00
#> (Intercept):63 -1.80051762 0.45360538  -3.9693480 7.206956e-05
#> (Intercept):64 -2.05992993 0.45567478  -4.5206143 6.166045e-06
#> (Intercept):65  2.25551661 0.20905908  10.7888959 0.000000e+00
#> (Intercept):66  0.55619490 0.41025099   1.3557430 1.751810e-01
#> (Intercept):67  2.70009438 0.20578113  13.1211953 0.000000e+00
#> (Intercept):68  3.13129847 0.22177693  14.1191352 0.000000e+00
#> (Intercept):69  3.59352340 0.20368961  17.6421540 0.000000e+00
#> (Intercept):70  2.93091828 0.20402810  14.3652676 0.000000e+00
#> (Intercept):71  2.33191454 0.20662480  11.2857438 0.000000e+00
#> (Intercept):72  1.88141868 0.21083896   8.9234870 0.000000e+00
#> (Intercept):73  1.78114720 0.21440200   8.3075120 0.000000e+00
#> (Intercept):74  2.01368480 0.20891750   9.6386604 0.000000e+00
#> (Intercept):75  2.16799098 0.20927509  10.3595272 0.000000e+00
#> (Intercept):76  2.96472763 0.20760851  14.2803760 0.000000e+00
#> (Intercept):77  2.56744774 0.20614939  12.4543068 0.000000e+00
#> (Intercept):78  2.39641394 0.21364951  11.2165667 0.000000e+00
#> (Intercept):79  1.57184389 0.21367232   7.3563291 1.889600e-13
#> (Intercept):80  0.34731915 0.24650871   1.4089529 1.588491e-01
#> (Intercept):81  1.82239037 0.31814478   5.7281795 1.015141e-08
#> (Intercept):82  2.27708630 0.20772302  10.9621279 0.000000e+00
#> (Intercept):83  2.27342827 0.21207431  10.7199605 0.000000e+00
#> (Intercept):84  2.14224808 0.20637172  10.3805311 0.000000e+00
#> (Intercept):85  3.15186577 0.20760356  15.1821376 0.000000e+00
#> (Intercept):86  2.66301057 0.27795496   9.5807268 0.000000e+00
#> (Intercept):87 -0.43164908 0.27781951  -1.5537032 1.202552e-01
#> (Intercept):88 -1.07987315 0.33058186  -3.2665833 1.088538e-03
#> (Intercept):89  2.74901382 0.20783448  13.2269380 0.000000e+00
#> (Intercept):90  2.38360223 0.20805844  11.4564071 0.000000e+00
#> (Intercept):91  0.54475302 0.26194172   2.0796726 3.755557e-02
#> (Intercept):92  3.15410012 0.21033948  14.9952832 0.000000e+00
#> (Intercept):93  1.88618892 0.21354085   8.8329185 0.000000e+00
#> (Intercept):94 -1.80040922 0.73523997  -2.4487369 1.433581e-02
#> yearsold_1      0.20367964 0.03017610   6.7496998 1.481504e-11
#> yearsold_2     -0.08609727 0.03729576  -2.3085002 2.097133e-02
#> yearsold_3     -0.32985926 0.04742935  -6.9547500 3.531841e-12
#> yearsold_4     -0.62441954 0.05932327 -10.5257108 0.000000e+00
#> yearsold_5     -0.78997892 0.07262616 -10.8773333 0.000000e+00
#> yearsold_6     -0.87753831 0.08417934 -10.4246284 0.000000e+00
#> yearsold_7     -0.97669407 0.09826994  -9.9388896 0.000000e+00
#> yearsold_8     -0.82061006 0.12647660  -6.4882360 8.684720e-11