Estimated coefficients and related statistics from the model ml.Repeatr4

Basic choice model

results_ml_Repeatr4

Format

dataframe with one row for each coefficient in the model.

Estimate

The coefficient value

Std. Error

The standard error of the coefficient

z-value

The z-value of the coefficient

Pr(>|z|)

The P value of the coefficient

Details

This model was estimated with mlogit on all of the data.

The utility formula was as follows:

choice ~ yearsold_1 + yearsold_2 + yearsold_3 + yearsold_4 + yearsold_5 + yearsold_6 + yearsold_7 + yearsold_8 + song2 + ... + song92

Examples

  results_ml_Repeatr4
#>                   Estimate Std. Error     z-value     Pr(>|z|)
#> (Intercept):2   2.54701258 0.21280308  11.9688710 0.000000e+00
#> (Intercept):3   2.69853425 0.24621944  10.9598750 0.000000e+00
#> (Intercept):4   2.87779047 0.22046117  13.0535028 0.000000e+00
#> (Intercept):5   2.20134903 0.21648519  10.1685895 0.000000e+00
#> (Intercept):6   2.18727991 0.21413660  10.2144140 0.000000e+00
#> (Intercept):7   3.61334126 0.20723645  17.4358381 0.000000e+00
#> (Intercept):8   2.78458703 0.21044900  13.2316480 0.000000e+00
#> (Intercept):9   3.01315076 0.20405963  14.7660306 0.000000e+00
#> (Intercept):10  3.61570511 0.21758134  16.6177170 0.000000e+00
#> (Intercept):11  1.21955471 0.22311493   5.4660381 4.602055e-08
#> (Intercept):12  1.46967100 0.21688285   6.7763357 1.232614e-11
#> (Intercept):13  1.89926950 0.21398010   8.8759166 0.000000e+00
#> (Intercept):14  1.80130947 0.21609923   8.3355664 0.000000e+00
#> (Intercept):15  1.82086120 0.21424807   8.4988454 0.000000e+00
#> (Intercept):16  2.57936659 0.20705160  12.4576026 0.000000e+00
#> (Intercept):17  3.32265139 0.25353739  13.1051731 0.000000e+00
#> (Intercept):18  1.81999435 0.21138123   8.6100094 0.000000e+00
#> (Intercept):19  1.57516013 0.24538520   6.4191327 1.370528e-10
#> (Intercept):20  3.33333621 0.22061629  15.1092023 0.000000e+00
#> (Intercept):21  0.11425982 0.29510589   0.3871824 6.986211e-01
#> (Intercept):22  1.60841970 0.21421941   7.5082819 5.995204e-14
#> (Intercept):23  3.43871483 0.20798230  16.5336898 0.000000e+00
#> (Intercept):24  1.90550167 0.22018946   8.6539188 0.000000e+00
#> (Intercept):25  2.28730921 0.28258029   8.0943691 6.661338e-16
#> (Intercept):26  2.39098251 0.25942811   9.2163585 0.000000e+00
#> (Intercept):27  2.17271864 0.20767724  10.4619968 0.000000e+00
#> (Intercept):28  2.78095406 0.20559875  13.5261232 0.000000e+00
#> (Intercept):29  2.61699929 0.22527665  11.6168247 0.000000e+00
#> (Intercept):30  1.89047929 0.21433273   8.8203016 0.000000e+00
#> (Intercept):31  2.56055444 0.22464700  11.3981245 0.000000e+00
#> (Intercept):32  2.53528069 0.22541860  11.2469896 0.000000e+00
#> (Intercept):33  1.26393552 0.27166786   4.6525029 3.279302e-06
#> (Intercept):34  2.60825845 0.21027221  12.4041996 0.000000e+00
#> (Intercept):35  2.36181572 0.29794989   7.9268890 2.220446e-15
#> (Intercept):36  0.86853668 0.23082264   3.7627881 1.680296e-04
#> (Intercept):37  2.38155170 0.21106401  11.2835517 0.000000e+00
#> (Intercept):38  1.09750592 0.22172257   4.9499061 7.424930e-07
#> (Intercept):39  2.41039706 0.20601141  11.7003087 0.000000e+00
#> (Intercept):40  1.72646074 0.21454672   8.0470155 8.881784e-16
#> (Intercept):41  1.15024574 0.26923607   4.2722572 1.935042e-05
#> (Intercept):42 -0.28557119 0.74090421  -0.3854361 6.999144e-01
#> (Intercept):43 -0.56057228 0.28745627  -1.9501133 5.116262e-02
#> (Intercept):44  2.50097683 0.20592453  12.1451139 0.000000e+00
#> (Intercept):45  1.17266555 0.22435554   5.2268178 1.724523e-07
#> (Intercept):46  1.40565402 0.21963881   6.3998435 1.555363e-10
#> (Intercept):47  1.50659936 0.21468742   7.0176416 2.256417e-12
#> (Intercept):48  0.77930402 0.25518615   3.0538649 2.259138e-03
#> (Intercept):49  1.76047209 0.21169755   8.3159775 0.000000e+00
#> (Intercept):50  2.41184585 0.31143050   7.7444112 9.547918e-15
#> (Intercept):51  1.46840552 0.21952249   6.6890891 2.245648e-11
#> (Intercept):52  1.84883761 0.21982428   8.4105251 0.000000e+00
#> (Intercept):53  2.70579304 0.20667260  13.0921714 0.000000e+00
#> (Intercept):54  2.78001580 0.20869859  13.3207214 0.000000e+00
#> (Intercept):55  2.95810348 0.21144608  13.9898712 0.000000e+00
#> (Intercept):56  1.30928385 0.21868139   5.9871754 2.135164e-09
#> (Intercept):57  2.16211755 0.26132842   8.2735645 2.220446e-16
#> (Intercept):58  2.26697062 0.22922079   9.8898996 0.000000e+00
#> (Intercept):59  3.02276304 0.23146518  13.0592558 0.000000e+00
#> (Intercept):60  2.68115226 0.23904448  11.2161227 0.000000e+00
#> (Intercept):61  1.80699951 0.23579211   7.6635285 1.798561e-14
#> (Intercept):62  3.22893258 0.22114688  14.6008510 0.000000e+00
#> (Intercept):63 -1.80116432 0.45360523  -3.9707750 7.163920e-05
#> (Intercept):64 -2.06153649 0.45567391  -4.5241486 6.063910e-06
#> (Intercept):65  2.25387355 0.20905709  10.7811388 0.000000e+00
#> (Intercept):66  0.55442619 0.41025020   1.3514343 1.765563e-01
#> (Intercept):67  2.70025987 0.20578109  13.1220018 0.000000e+00
#> (Intercept):68  3.13290576 0.22177522  14.1264913 0.000000e+00
#> (Intercept):69  3.59244412 0.20368849  17.6369517 0.000000e+00
#> (Intercept):70  2.93041346 0.20402784  14.3628118 0.000000e+00
#> (Intercept):71  2.33062934 0.20662331  11.2796053 0.000000e+00
#> (Intercept):72  1.87987470 0.21083710   8.9162426 0.000000e+00
#> (Intercept):73  1.78112951 0.21440198   8.3074302 0.000000e+00
#> (Intercept):74  2.01246285 0.20891627   9.6328680 0.000000e+00
#> (Intercept):75  2.16647985 0.20927324  10.3523980 0.000000e+00
#> (Intercept):76  2.96317960 0.20760661  14.2730501 0.000000e+00
#> (Intercept):77  2.56742743 0.20614936  12.4542102 0.000000e+00
#> (Intercept):78  2.39507655 0.21364697  11.2104402 0.000000e+00
#> (Intercept):79  1.57117131 0.21367201   7.3531920 1.936229e-13
#> (Intercept):80  0.34669991 0.24650847   1.4064422 1.595929e-01
#> (Intercept):81  1.82503243 0.31814262   5.7365229 9.663998e-09
#> (Intercept):82  2.27599413 0.20772214  10.9569162 0.000000e+00
#> (Intercept):83  2.27224781 0.21207117  10.7145529 0.000000e+00
#> (Intercept):84  2.14194128 0.20637163  10.3790491 0.000000e+00
#> (Intercept):85  3.15236674 0.20760306  15.1845873 0.000000e+00
#> (Intercept):86  2.66554718 0.27795180   9.5899620 0.000000e+00
#> (Intercept):87 -0.43323950 0.27781786  -1.5594372 1.188929e-01
#> (Intercept):88 -1.15083178 0.33770595  -3.4077924 6.549070e-04
#> (Intercept):89  2.74757872 0.20783261  13.2201520 0.000000e+00
#> (Intercept):90  2.38208971 0.20805669  11.4492340 0.000000e+00
#> (Intercept):91  0.54540344 0.26194144   2.0821579 3.732805e-02
#> (Intercept):92  3.15295118 0.21033727  14.9899787 0.000000e+00
#> (Intercept):93  1.88617607 0.21354083   8.8328591 0.000000e+00
#> (Intercept):94 -1.79941906 0.73523965  -2.4473912 1.438946e-02
#> yearsold_1      0.20393718 0.03017557   6.7583546 1.395684e-11
#> yearsold_2     -0.08585554 0.03729477  -2.3020797 2.133068e-02
#> yearsold_3     -0.32831432 0.04742280  -6.9231319 4.417577e-12
#> yearsold_4     -0.62301205 0.05931798 -10.5029213 0.000000e+00
#> yearsold_5     -0.78814430 0.07261895 -10.8531490 0.000000e+00
#> yearsold_6     -0.87537051 0.08417129 -10.3998709 0.000000e+00
#> yearsold_7     -0.97420430 0.09825959  -9.9145975 0.000000e+00
#> yearsold_8     -0.81663808 0.12645168  -6.4581038 1.060230e-10