Estimated coefficients and related statistics from the model ml.Repeatr4

Basic choice model

results_ml_Repeatr4

Format

dataframe with one row for each coefficient in the model.

Estimate

The coefficient value

Std. Error

The standard error of the coefficient

z-value

The z-value of the coefficient

Pr(>|z|)

The P value of the coefficient

Details

This model was estimated with mlogit on all of the data.

The utility formula was as follows:

choice ~ yearsold_1 + yearsold_2 + yearsold_3 + yearsold_4 + yearsold_5 + yearsold_6 + yearsold_7 + yearsold_8 + song2 + ... + song92

Examples

  results_ml_Repeatr4
#>                   Estimate Std. Error      z-value     Pr(>|z|)
#> (Intercept):2   0.45030639 0.06658980   6.76239318 1.357314e-11
#> (Intercept):3  -0.12680104 0.06525454  -1.94317573 5.199493e-02
#> (Intercept):4  -0.44175014 0.06428819  -6.87140448 6.357359e-12
#> (Intercept):5  -0.19197359 0.06243947  -3.07455525 2.108167e-03
#> (Intercept):6  -0.37551004 0.06513180  -5.76538686 8.147074e-09
#> (Intercept):7  -0.17975804 0.06732422  -2.67003520 7.584330e-03
#> (Intercept):8  -0.81652121 0.06902428 -11.82947821 0.000000e+00
#> (Intercept):9  -0.40112951 0.06840641  -5.86391739 4.520724e-09
#> (Intercept):10 -0.89094136 0.06809453 -13.08389099 0.000000e+00
#> (Intercept):11 -0.58252352 0.06703982  -8.68921625 0.000000e+00
#> (Intercept):12 -0.76584901 0.06966327 -10.99358424 0.000000e+00
#> (Intercept):13 -0.76515342 0.06736044 -11.35909132 0.000000e+00
#> (Intercept):14 -0.20806712 0.08652340  -2.40474964 1.618355e-02
#> (Intercept):15 -0.87123644 0.06875364 -12.67185992 0.000000e+00
#> (Intercept):16 -0.98300886 0.07229117 -13.59790989 0.000000e+00
#> (Intercept):17 -0.75945046 0.06996525 -10.85468069 0.000000e+00
#> (Intercept):18  0.46859484 0.11654996   4.02054927 5.806260e-05
#> (Intercept):19 -0.36681033 0.08625306  -4.25272226 2.111874e-05
#> (Intercept):20 -0.86292190 0.07901545 -10.92092558 0.000000e+00
#> (Intercept):21 -0.96852724 0.07334394 -13.20527899 0.000000e+00
#> (Intercept):22 -0.96871533 0.08136545 -11.90573372 0.000000e+00
#> (Intercept):23 -1.26371452 0.07548408 -16.74147049 0.000000e+00
#> (Intercept):24  0.29304699 0.11808669   2.48162593 1.307845e-02
#> (Intercept):25 -0.64451832 0.09115428  -7.07063122 1.542322e-12
#> (Intercept):26 -0.73124539 0.09121110  -8.01706598 1.110223e-15
#> (Intercept):27 -1.13531446 0.07971793 -14.24164511 0.000000e+00
#> (Intercept):28 -0.45294379 0.10020051  -4.52037430 6.173039e-06
#> (Intercept):29  0.01420683 0.11485248   0.12369632 9.015557e-01
#> (Intercept):30 -0.58085450 0.09912076  -5.86006889 4.626752e-09
#> (Intercept):31 -1.00934098 0.09539338 -10.58082860 0.000000e+00
#> (Intercept):32 -0.57041575 0.10374280  -5.49836453 3.833299e-08
#> (Intercept):33 -1.25538962 0.07971469 -15.74853477 0.000000e+00
#> (Intercept):34 -1.35057955 0.08059805 -16.75697589 0.000000e+00
#> (Intercept):35 -1.33309142 0.08131233 -16.39470282 0.000000e+00
#> (Intercept):36 -0.53670908 0.12009937  -4.46887521 7.863199e-06
#> (Intercept):37 -0.34920887 0.12021965  -2.90475697 3.675384e-03
#> (Intercept):38 -1.42368547 0.08648533 -16.46158360 0.000000e+00
#> (Intercept):39 -1.38648975 0.09322195 -14.87299589 0.000000e+00
#> (Intercept):40 -1.73185891 0.08685698 -19.93920134 0.000000e+00
#> (Intercept):41 -1.32146639 0.09965287 -13.26069634 0.000000e+00
#> (Intercept):42  0.44723617 0.14373125   3.11161402 1.860676e-03
#> (Intercept):43 -1.68883878 0.08914705 -18.94441644 0.000000e+00
#> (Intercept):44  0.17909670 0.15191130   1.17895574 2.384158e-01
#> (Intercept):45 -1.69142297 0.09246895 -18.29179452 0.000000e+00
#> (Intercept):46 -0.26812581 0.15111157  -1.77435659 7.600419e-02
#> (Intercept):47 -1.57041292 0.09989197 -15.72111289 0.000000e+00
#> (Intercept):48 -1.53397125 0.10073920 -15.22715303 0.000000e+00
#> (Intercept):49 -0.01150394 0.15252109  -0.07542524 9.398764e-01
#> (Intercept):50 -1.25648808 0.11371802 -11.04915571 0.000000e+00
#> (Intercept):51 -1.63468831 0.10658453 -15.33701233 0.000000e+00
#> (Intercept):52 -0.95374308 0.13338918  -7.15007828 8.673062e-13
#> (Intercept):53 -1.92330878 0.09585717 -20.06431870 0.000000e+00
#> (Intercept):54 -1.27765497 0.12081243 -10.57552537 0.000000e+00
#> (Intercept):55 -1.36228980 0.11520340 -11.82508310 0.000000e+00
#> (Intercept):56  0.07245640 0.14861960   0.48752924 6.258833e-01
#> (Intercept):57 -2.05036021 0.09692964 -21.15307718 0.000000e+00
#> (Intercept):58 -1.96769015 0.09757692 -20.16552771 0.000000e+00
#> (Intercept):59 -0.52746118 0.15781153  -3.34234889 8.307257e-04
#> (Intercept):60 -1.84343656 0.11048773 -16.68453643 0.000000e+00
#> (Intercept):61 -1.30766708 0.13882288  -9.41967955 0.000000e+00
#> (Intercept):62 -1.23552324 0.13904246  -8.88594176 0.000000e+00
#> (Intercept):63 -0.57656177 0.15384235  -3.74774408 1.784322e-04
#> (Intercept):64 -0.12677862 0.17343288  -0.73099530 4.647820e-01
#> (Intercept):65 -0.60029852 0.15535603  -3.86401814 1.115369e-04
#> (Intercept):66 -2.28160729 0.11175468 -20.41621331 0.000000e+00
#> (Intercept):67 -0.88580050 0.16086552  -5.50646573 3.661092e-08
#> (Intercept):68 -0.48306754 0.18445614  -2.61887490 8.822029e-03
#> (Intercept):69 -1.34556466 0.16778827  -8.01942026 1.110223e-15
#> (Intercept):70 -0.43399497 0.19425195  -2.23418588 2.547084e-02
#> (Intercept):71  0.13209223 0.20817311   0.63453069 5.257345e-01
#> (Intercept):72 -1.57727792 0.18063499  -8.73185168 0.000000e+00
#> (Intercept):73 -0.74036765 0.21079055  -3.51233798 4.441827e-04
#> (Intercept):74 -2.79511903 0.15910109 -17.56819540 0.000000e+00
#> (Intercept):75 -1.00182710 0.21458881  -4.66858972 3.032743e-06
#> (Intercept):76 -2.36747574 0.18973778 -12.47761896 0.000000e+00
#> (Intercept):77 -0.44516147 0.23755177  -1.87395561 6.093654e-02
#> (Intercept):78 -2.59916441 0.19724366 -13.17742970 0.000000e+00
#> (Intercept):79 -1.87857680 0.22245140  -8.44488634 0.000000e+00
#> (Intercept):80 -2.00281303 0.21389299  -9.36362173 0.000000e+00
#> (Intercept):81 -3.53497877 0.18750330 -18.85288845 0.000000e+00
#> (Intercept):82 -0.86881068 0.24388433  -3.56238830 3.674963e-04
#> (Intercept):83 -3.14317843 0.21045632 -14.93506301 0.000000e+00
#> (Intercept):84 -3.69616928 0.20433978 -18.08834887 0.000000e+00
#> (Intercept):85 -0.81449594 0.26377944  -3.08779155 2.016499e-03
#> (Intercept):86 -3.03124558 0.24079049 -12.58872633 0.000000e+00
#> (Intercept):87 -0.76678281 0.28055143  -2.73312749 6.273604e-03
#> (Intercept):88 -1.29331238 0.28363753  -4.55973643 5.121786e-06
#> (Intercept):89 -4.21526257 0.26154398 -16.11684004 0.000000e+00
#> (Intercept):90 -2.55289967 0.35906316  -7.10989022 1.161293e-12
#> (Intercept):91 -4.94324764 0.41247802 -11.98426921 0.000000e+00
#> (Intercept):92 -3.41225261 0.72679796  -4.69491219 2.667213e-06
#> yearsold_1      0.21307040 0.03027868   7.03697682 1.964651e-12
#> yearsold_2     -0.08061738 0.03741235  -2.15483340 3.117487e-02
#> yearsold_3     -0.32183489 0.04761389  -6.75926543 1.386935e-11
#> yearsold_4     -0.61488737 0.05952156 -10.33049904 0.000000e+00
#> yearsold_5     -0.77924027 0.07288550 -10.69129350 0.000000e+00
#> yearsold_6     -0.86741641 0.08457893 -10.25570297 0.000000e+00
#> yearsold_7     -0.96129024 0.09871895  -9.73764651 0.000000e+00
#> yearsold_8     -0.81569544 0.12687182  -6.42928786 1.282032e-10