Fugazi songs data

Song data from the Fugazi discography pages on Wikipedia. The variables attributing lead vocals are simplifications in some cases where lead vocals were shared.

songvarslookup

Format

dataframe with one row for each song in the Fugazi discography, except those which never appear in the Fugazi Live Series data.

songid

numeric id for each song

releaseid

numeric id in ascending chronological order

release

name of album or EP

track_number

The track number for the song on the release

instrumental

Indicates whether or not the piece is an instrumental

vocals_picciotto

indicates whether or not Guy Picciotto sang lead vocals on this track

vocals_mackaye

indicates whether or not Ian Mackaye sang lead vocals on this track

vocals_lally

indicates whether or not Joe Lally sang lead vocals on this track

duration_seconds

The duration of the song in seconds

Source

https://web.archive.org/web/20201112000517/http://en.wikipedia.org/wiki/Fugazi_discography

Examples

  songvarslookup
#>    rank_length releaseid             release track_number songid
#> 1            1         6 In on the Killtaker            6     83
#> 2            2         9        The Argument            7     88
#> 3            3         8            End Hits            8     65
#> 4            4         9        The Argument            6     77
#> 5            5         7        Red Medicine            8     32
#> 6            6         8            End Hits            7     44
#> 7            7         4            Repeater           11     16
#> 8            8         1              Fugazi            6     15
#> 9            9         6 In on the Killtaker           12     51
#> 10          10        11          First Demo            5     81
#> 11          11         9        The Argument            8     68
#> 12          12         8            End Hits           10     46
#> 13          13         9        The Argument           11     70
#> 14          14         9        The Argument            2     71
#> 15          15         1              Fugazi            7     57
#> 16          16         8            End Hits            9     79
#> 17          17         9        The Argument            9     73
#> 18          18         7        Red Medicine           13     61
#> 19          19         4            Repeater            1      9
#> 20          20         8            End Hits            4     67
#> 21          21         8            End Hits           12     69
#> 22          22         2       Margin Walker            6     10
#> 23          23         9        The Argument           10     75
#> 24          24         6 In on the Killtaker            4     30
#> 25          25         9        The Argument            4     82
#> 26          26         5         Steady Diet            8     27
#> 27          27         9        The Argument            3     85
#> 28          28         4            Repeater            5      3
#> 29          29         8            End Hits            3     49
#> 30          30         6 In on the Killtaker            5     14
#> 31          31         7        Red Medicine            7     54
#> 32          32         6 In on the Killtaker           11     25
#> 33          33        11          First Demo            8     89
#> 34          34         5         Steady Diet            6     74
#> 35          35         8            End Hits           13     59
#> 36          36         5         Steady Diet            9     91
#> 37          37         6 In on the Killtaker            7     31
#> 38          38        10           Furniture            1     66
#> 39          39         7        Red Medicine            3     76
#> 40          40         7        Red Medicine           10     29
#> 41          41         4            Repeater            8     12
#> 42          42         2       Margin Walker            2     11
#> 43          43         5         Steady Diet           10     48
#> 44          44         5         Steady Diet            3     60
#> 45          45         4            Repeater            6      7
#> 46          46         5         Steady Diet            2      2
#> 47          47         7        Red Medicine            9     78
#> 48          48         6 In on the Killtaker           10     50
#> 49          49         7        Red Medicine            1     24
#> 50          50         5         Steady Diet            5     39
#> 51          51         6 In on the Killtaker            3     55
#> 52          52         5         Steady Diet            1     22
#> 53          53         9        The Argument            5     87
#> 54          54        10           Furniture            2     64
#> 55          55         5         Steady Diet            4     47
#> 56          56         7        Red Medicine            4     37
#> 57          57         7        Red Medicine            6     86
#> 58          58         7        Red Medicine            5     36
#> 59          59         3             3 Songs            2     58
#> 60          60         4            Repeater            2      8
#> 61          61         4            Repeater            4      5
#> 62          62         1              Fugazi            5     13
#> 63          63         5         Steady Diet           11     40
#> 64          64         8            End Hits           11     80
#> 65          65         3             3 Songs            1     17
#> 66          66         1              Fugazi            1      1
#> 67          67         1              Fugazi            2     33
#> 68          68         7        Red Medicine           12     62
#> 69          69         2       Margin Walker            3     35
#> 70          70         7        Red Medicine            2     18
#> 71          71        11          First Demo           10     84
#> 72          72         8            End Hits            2     56
#> 73          73         6 In on the Killtaker            1     19
#> 74          74         1              Fugazi            4     34
#> 75          75         1              Fugazi            3     21
#> 76          76         4            Repeater            9     20
#> 77          77         4            Repeater            3     45
#> 78          78         2       Margin Walker            1      6
#> 79          79         6 In on the Killtaker            8     41
#> 80          80         8            End Hits            5     63
#> 81          81         4            Repeater           10     23
#> 82          82         2       Margin Walker            4     90
#> 83          83         5         Steady Diet            7      4
#> 84          84         8            End Hits            1     42
#> 85          85         2       Margin Walker            5     43
#> 86          86        10           Furniture            3     93
#> 87          87         6 In on the Killtaker            2     28
#> 88          88         8            End Hits            6     72
#> 89          89         6 In on the Killtaker            9     26
#> 90          90         4            Repeater            7     38
#> 91          91         7        Red Medicine           11     52
#> 92          92         3             3 Songs            3     53
#> 93          93         0          Unreleased            1     92
#>                            song instrumental vocals_picciotto vocals_mackaye
#> 1                  23 beats off            0                0              1
#> 2                  strangelight            0                1              0
#> 3                  floating boy            0                1              0
#> 4                      the kill            0                0              0
#> 5                        by you            0                0              0
#> 6              closed captioned            0                0              1
#> 7                 shut the door            0                0              1
#> 8                    suggestion            0                0              1
#> 9  last chance for a slow dance            0                1              0
#> 10                     the word            0                0              1
#> 11                           oh            0                1              0
#> 12                  arpeggiator            1                0              0
#> 13                     argument            0                0              1
#> 14                      cashout            0                0              1
#> 15                      glueman            0                1              0
#> 16                foreman's dog            0                1              0
#> 17                 ex-spectator            0                0              1
#> 18         long distance runner            0                0              1
#> 19                     turnover            0                1              0
#> 20                  no surprise            0                1              0
#> 21                  pink frosty            0                0              1
#> 22                     promises            0                0              1
#> 23                    nightshop            0                1              0
#> 24            smallpox champion            0                1              0
#> 25                 epic problem            0                0              1
#> 26               runaway return            0                1              0
#> 27              full disclosure            0                1              0
#> 28                    blueprint            0                1              0
#> 29                recap modotti            0                0              0
#> 30                      rend it            0                1              0
#> 31              fell, destroyed            0                1              0
#> 32                   instrument            0                0              1
#> 33           turn off your guns            0                0              1
#> 34                  steady diet            1                0              0
#> 35                           fd            0                0              1
#> 36                       polish            0                0              1
#> 37                sweet and low            1                0              0
#> 38                    furniture            0                0              1
#> 39              latest disgrace            0                1              0
#> 40                       target            0                1              0
#> 41                two beats off            0                1              0
#> 42                 and the same            0                0              1
#> 43          dear justice letter            0                1              0
#> 44              nice new outfit            0                1              0
#> 45            sieve-fisted find            0                1              0
#> 46                  reclamation            0                0              1
#> 47                      version            1                0              0
#> 48            walken's syndrome            0                1              0
#> 49               do you like me            0                1              0
#> 50                  latin roots            0                1              0
#> 51          returning the screw            0                0              1
#> 52                    exit only            0                1              0
#> 53                life and limb            0                1              0
#> 54                     number 5            1                0              0
#> 55                       stacks            0                0              1
#> 56                birthday pony            0                0              1
#> 57             combination lock            1                0              0
#> 58               forensic scene            0                1              0
#> 59                       joe #1            1                0              0
#> 60                     repeater            0                0              1
#> 61                  merchandise            0                0              1
#> 62             give me the cure            0                1              0
#> 63                         kyeo            0                0              1
#> 64                guilford fall            0                1              0
#> 65                      song #1            0                0              1
#> 66                 waiting room            0                0              1
#> 67                bulldog front            0                1              0
#> 68                  downed city            0                1              0
#> 69                  burning too            0                0              1
#> 70         bed for the scraping            0                0              1
#> 71         in defense of humans            0                0              1
#> 72               place position            0                1              0
#> 73                facet squared            0                0              1
#> 74                      burning            0                1              0
#> 75                    bad mouth            0                0              1
#> 76                    styrofoam            0                0              1
#> 77                   brendan #1            1                0              0
#> 78                margin walker            0                1              0
#> 79                   cassavetes            0                1              0
#> 80            five corporations            0                0              1
#> 81                reprovisional            0                1              0
#> 82                  provisional            0                1              0
#> 83                long division            0                0              1
#> 84                        break            0                0              1
#> 85                     lockdown            0                1              0
#> 86                hello morning            0                1              0
#> 87       public witness program            0                1              0
#> 88             caustic acrostic            0                1              0
#> 89                    great cop            0                0              1
#> 90                        greed            0                0              1
#> 91                 back to base            0                0              1
#> 92                     break-in            0                1              0
#> 93               preprovisional            0                1              0
#>    vocals_lally duration_seconds
#> 1             0              401
#> 2             0              353
#> 3             0              345
#> 4             1              327
#> 5             1              311
#> 6             0              292
#> 7             0              289
#> 8             0              284
#> 9             0              278
#> 10            0              278
#> 11            0              269
#> 12            0              268
#> 13            0              267
#> 14            0              264
#> 15            0              263
#> 16            0              261
#> 17            0              258
#> 18            0              257
#> 19            0              256
#> 20            0              252
#> 21            0              249
#> 22            0              242
#> 23            0              242
#> 24            0              241
#> 25            0              239
#> 26            0              238
#> 27            0              233
#> 28            0              232
#> 29            1              230
#> 30            0              228
#> 31            0              226
#> 32            0              223
#> 33            0              223
#> 34            0              222
#> 35            0              222
#> 36            0              218
#> 37            0              216
#> 38            0              215
#> 39            0              214
#> 40            0              212
#> 41            0              208
#> 42            0              207
#> 43            0              207
#> 44            0              206
#> 45            0              204
#> 46            0              201
#> 47            0              200
#> 48            0              198
#> 49            0              196
#> 50            0              193
#> 51            0              193
#> 52            0              191
#> 53            0              189
#> 54            0              189
#> 55            0              188
#> 56            0              188
#> 57            0              186
#> 58            0              185
#> 59            0              181
#> 60            0              181
#> 61            0              179
#> 62            0              178
#> 63            0              178
#> 64            0              177
#> 65            0              174
#> 66            0              173
#> 67            0              173
#> 68            0              173
#> 69            0              170
#> 70            0              170
#> 71            0              167
#> 72            0              165
#> 73            0              162
#> 74            0              159
#> 75            0              155
#> 76            0              154
#> 77            0              152
#> 78            0              150
#> 79            0              150
#> 80            0              149
#> 81            0              138
#> 82            0              137
#> 83            0              132
#> 84            0              132
#> 85            0              130
#> 86            0              126
#> 87            0              124
#> 88            0              121
#> 89            0              112
#> 90            0              107
#> 91            0              105
#> 92            0               92
#> 93            0              130