Fugazi songs data

Song data from the Fugazi discography pages on Wikipedia. The variables attributing lead vocals are simplifications in some cases where lead vocals were shared.

songvarslookup

Format

dataframe with one row for each song in the Fugazi discography, except those which never appear in the Fugazi Live Series data.

songid

numeric id for each song

releaseid

numeric id in ascending chronological order

track_number

The track number for the song on the release

instrumental

Indicates whether or not the piece is an instrumental

vocals_picciotto

indicates whether or not Guy Picciotto sang lead vocals on this track

vocals_mackaye

indicates whether or not Ian Mackaye sang lead vocals on this track

vocals_lally

indicates whether or not Joe Lally sang lead vocals on this track

duration_seconds

The duration of the song in seconds

Source

https://web.archive.org/web/20201112000517/http://en.wikipedia.org/wiki/Fugazi_discography

Examples

  songvarslookup
#>    rank_length releaseid track_number                         song songid
#> 1           93         0            1               preprovisional     64
#> 2           66         1            1                 waiting room     92
#> 3           67         1            2                bulldog front     13
#> 4           75         1            3                    bad mouth      6
#> 5           74         1            4                      burning     14
#> 6           62         1            5             give me the cure     37
#> 7            8         1            6                   suggestion     83
#> 8           15         1            7                      glueman     38
#> 9           78         2            1                margin walker     54
#> 10          42         2            2                 and the same      2
#> 11          69         2            3                  burning too     15
#> 12          82         2            4                  provisional     66
#> 13          85         2            5                     lockdown     51
#> 14          22         2            6                     promises     65
#> 15          65         3            1                      song #1     78
#> 16          59         3            2                       joe #1     45
#> 17          92         3            3                     break-in     11
#> 18          19         4            1                     turnover     89
#> 19          60         4            2                     repeater     71
#> 20          77         4            3                   brendan #1     12
#> 21          61         4            4                  merchandise     55
#> 22          28         4            5                    blueprint      9
#> 23          45         4            6            sieve-fisted find     76
#> 24          90         4            7                        greed     40
#> 25          41         4            8                two beats off     90
#> 26          76         4            9                    styrofoam     82
#> 27          81         4           10                reprovisional     72
#> 28           7         4           11                shut the door     75
#> 29          52         5            1                    exit only     27
#> 30          46         5            2                  reclamation     69
#> 31          44         5            3              nice new outfit     56
#> 32          55         5            4                       stacks     79
#> 33          50         5            5                  latin roots     49
#> 34          34         5            6                  steady diet     80
#> 35          83         5            7                long division     53
#> 36          26         5            8               runaway return     74
#> 37          36         5            9                       polish     63
#> 38          43         5           10          dear justice letter     22
#> 39          63         5           11                         kyeo     46
#> 40          73         6            1                facet squared     28
#> 41          87         6            2       public witness program     67
#> 42          51         6            3          returning the screw     73
#> 43          24         6            4            smallpox champion     77
#> 44          30         6            5                      rend it     70
#> 45           1         6            6                 23 beats off      1
#> 46          37         6            7                sweet and low     84
#> 47          79         6            8                   cassavetes     18
#> 48          89         6            9                    great cop     39
#> 49          48         6           10            walken's syndrome     93
#> 50          32         6           11                   instrument     44
#> 51           9         6           12 last chance for a slow dance     47
#> 52          49         7            1               do you like me     23
#> 53          70         7            2         bed for the scraping      7
#> 54          39         7            3              latest disgrace     48
#> 55          56         7            4                birthday pony      8
#> 56          58         7            5               forensic scene     34
#> 57          57         7            6             combination lock     21
#> 58          31         7            7              fell, destroyed     30
#> 59           5         7            8                       by you     16
#> 60          47         7            9                      version     91
#> 61          40         7           10                       target     85
#> 62          91         7           11                 back to base      5
#> 63          68         7           12                  downed city     24
#> 64          18         7           13         long distance runner     52
#> 65          84         8            1                        break     10
#> 66          72         8            2               place position     62
#> 67          29         8            3                recap modotti     68
#> 68          20         8            4                  no surprise     58
#> 69          80         8            5            five corporations     31
#> 70          88         8            6             caustic acrostic     19
#> 71           6         8            7             closed captioned     20
#> 72           3         8            8                 floating boy     32
#> 73          16         8            9                foreman's dog     33
#> 74          12         8           10                  arpeggiator      4
#> 75          64         8           11                guilford fall     41
#> 76          21         8           12                  pink frosty     61
#> 77          35         8           13                           fd     29
#> 78          14         9            2                      cashout     17
#> 79          27         9            3              full disclosure     35
#> 80          25         9            4                 epic problem     25
#> 81          53         9            5                life and limb     50
#> 82           4         9            6                     the kill     86
#> 83           2         9            7                 strangelight     81
#> 84          11         9            8                           oh     60
#> 85          17         9            9                 ex-spectator     26
#> 86          23         9           10                    nightshop     57
#> 87          13         9           11                     argument      3
#> 88          38        10            1                    furniture     36
#> 89          54        10            2                     number 5     59
#> 90          86        10            3                hello morning     42
#> 91          10        11            5                     the word     87
#> 92          33        11            8           turn off your guns     88
#> 93          71        11           10         in defense of humans     43
#>    instrumental vocals_picciotto vocals_mackaye vocals_lally duration_seconds
#> 1             0                1              0            0              130
#> 2             0                0              1            0              173
#> 3             0                1              0            0              173
#> 4             0                0              1            0              155
#> 5             0                1              0            0              159
#> 6             0                1              0            0              178
#> 7             0                0              1            0              284
#> 8             0                1              0            0              263
#> 9             0                1              0            0              150
#> 10            0                0              1            0              207
#> 11            0                0              1            0              170
#> 12            0                1              0            0              137
#> 13            0                1              0            0              130
#> 14            0                0              1            0              242
#> 15            0                0              1            0              174
#> 16            1                0              0            0              181
#> 17            0                1              0            0               92
#> 18            0                1              0            0              256
#> 19            0                0              1            0              181
#> 20            1                0              0            0              152
#> 21            0                0              1            0              179
#> 22            0                1              0            0              232
#> 23            0                1              0            0              204
#> 24            0                0              1            0              107
#> 25            0                1              0            0              208
#> 26            0                0              1            0              154
#> 27            0                1              0            0              138
#> 28            0                0              1            0              289
#> 29            0                1              0            0              191
#> 30            0                0              1            0              201
#> 31            0                1              0            0              206
#> 32            0                0              1            0              188
#> 33            0                1              0            0              193
#> 34            1                0              0            0              222
#> 35            0                0              1            0              132
#> 36            0                1              0            0              238
#> 37            0                0              1            0              218
#> 38            0                1              0            0              207
#> 39            0                0              1            0              178
#> 40            0                0              1            0              162
#> 41            0                1              0            0              124
#> 42            0                0              1            0              193
#> 43            0                1              0            0              241
#> 44            0                1              0            0              228
#> 45            0                0              1            0              401
#> 46            1                0              0            0              216
#> 47            0                1              0            0              150
#> 48            0                0              1            0              112
#> 49            0                1              0            0              198
#> 50            0                0              1            0              223
#> 51            0                1              0            0              278
#> 52            0                1              0            0              196
#> 53            0                0              1            0              170
#> 54            0                1              0            0              214
#> 55            0                0              1            0              188
#> 56            0                1              0            0              185
#> 57            1                0              0            0              186
#> 58            0                1              0            0              226
#> 59            0                0              0            1              311
#> 60            1                0              0            0              200
#> 61            0                1              0            0              212
#> 62            0                0              1            0              105
#> 63            0                1              0            0              173
#> 64            0                0              1            0              257
#> 65            0                0              1            0              132
#> 66            0                1              0            0              165
#> 67            0                0              0            1              230
#> 68            0                1              0            0              252
#> 69            0                0              1            0              149
#> 70            0                1              0            0              121
#> 71            0                0              1            0              292
#> 72            0                1              0            0              345
#> 73            0                1              0            0              261
#> 74            1                0              0            0              268
#> 75            0                1              0            0              177
#> 76            0                0              1            0              249
#> 77            0                0              1            0              222
#> 78            0                0              1            0              264
#> 79            0                1              0            0              233
#> 80            0                0              1            0              239
#> 81            0                1              0            0              189
#> 82            0                0              0            1              327
#> 83            0                1              0            0              353
#> 84            0                1              0            0              269
#> 85            0                0              1            0              258
#> 86            0                1              0            0              242
#> 87            0                0              1            0              267
#> 88            0                0              1            0              215
#> 89            1                0              0            0              189
#> 90            0                1              0            0              126
#> 91            0                0              1            0              278
#> 92            0                0              1            0              223
#> 93            0                0              1            0              167