Fugazi songs data

Song data from the Fugazi discography pages on Wikipedia. The variables attributing lead vocals are simplifications in some cases where lead vocals were shared.

songvarslookup

Format

dataframe with one row for each song in the Fugazi discography, except those which never appear in the Fugazi Live Series data.

songid

numeric id for each song

releaseid

numeric id in ascending chronological order

release

name of album or EP

track_number

The track number for the song on the release

instrumental

Indicates whether or not the piece is an instrumental

vocals_picciotto

indicates whether or not Guy Picciotto sang lead vocals on this track

vocals_mackaye

indicates whether or not Ian Mackaye sang lead vocals on this track

vocals_lally

indicates whether or not Joe Lally sang lead vocals on this track

duration_seconds

The duration of the song in seconds

Source

https://web.archive.org/web/20201112000517/http://en.wikipedia.org/wiki/Fugazi_discography

Examples

  songvarslookup
#>    rank_length releaseid track_number songid instrumental vocals_picciotto
#> 1            1         6            6     83            0                0
#> 2            2         9            7     88            0                1
#> 3            3         8            8     65            0                1
#> 4            4         9            6     77            0                0
#> 5            5         7            8     32            0                0
#> 6            6         8            7     44            0                0
#> 7            7         4           11     16            0                0
#> 8            8         1            6     15            0                0
#> 9            9         6           12     51            0                1
#> 10          10        11            5     81            0                0
#> 11          11         9            8     68            0                1
#> 12          12         8           10     46            1                0
#> 13          13         9           11     70            0                0
#> 14          14         9            2     71            0                0
#> 15          15         1            7     57            0                1
#> 16          16         8            9     79            0                1
#> 17          17         9            9     73            0                0
#> 18          18         7           13     61            0                0
#> 19          19         4            1      9            0                1
#> 20          20         8            4     67            0                1
#> 21          21         8           12     69            0                0
#> 22          22         2            6     10            0                0
#> 23          23         9           10     75            0                1
#> 24          24         6            4     30            0                1
#> 25          25         9            4     82            0                0
#> 26          26         5            8     27            0                1
#> 27          27         9            3     85            0                1
#> 28          28         4            5      3            0                1
#> 29          29         8            3     49            0                0
#> 30          30         6            5     14            0                1
#> 31          31         7            7     54            0                1
#> 32          32         6           11     25            0                0
#> 33          33        11            8     89            0                0
#> 34          34         5            6     74            1                0
#> 35          35         8           13     59            0                0
#> 36          36         5            9     91            0                0
#> 37          37         6            7     31            1                0
#> 38          38        10            1     66            0                0
#> 39          39         7            3     76            0                1
#> 40          40         7           10     29            0                1
#> 41          41         4            8     12            0                1
#> 42          42         2            2     11            0                0
#> 43          43         5           10     48            0                1
#> 44          44         5            3     60            0                1
#> 45          45         4            6      7            0                1
#> 46          46         5            2      2            0                0
#> 47          47         7            9     78            1                0
#> 48          48         6           10     50            0                1
#> 49          49         7            1     24            0                1
#> 50          50         5            5     39            0                1
#> 51          51         6            3     55            0                0
#> 52          52         5            1     22            0                1
#> 53          53         9            5     87            0                1
#> 54          54        10            2     64            1                0
#> 55          55         5            4     47            0                0
#> 56          56         7            4     37            0                0
#> 57          57         7            6     86            1                0
#> 58          58         7            5     36            0                1
#> 59          59         3            2     58            1                0
#> 60          60         4            2      8            0                0
#> 61          61         4            4      5            0                0
#> 62          62         1            5     13            0                1
#> 63          63         5           11     40            0                0
#> 64          64         8           11     80            0                1
#> 65          65         3            1     17            0                0
#> 66          66         1            1      1            0                0
#> 67          67         1            2     33            0                1
#> 68          68         7           12     62            0                1
#> 69          69         2            3     35            0                0
#> 70          70         7            2     18            0                0
#> 71          71        11           10     84            0                0
#> 72          72         8            2     56            0                1
#> 73          73         6            1     19            0                0
#> 74          74         1            4     34            0                1
#> 75          75         1            3     21            0                0
#> 76          76         4            9     20            0                0
#> 77          77         4            3     45            1                0
#> 78          78         2            1      6            0                1
#> 79          79         6            8     41            0                1
#> 80          80         8            5     63            0                0
#> 81          81         4           10     23            0                1
#> 82          82         2            4     90            0                1
#> 83          83         5            7      4            0                0
#> 84          84         8            1     42            0                0
#> 85          85         2            5     43            0                1
#> 86          86        10            3     93            0                1
#> 87          87         6            2     28            0                1
#> 88          88         8            6     72            0                1
#> 89          89         6            9     26            0                0
#> 90          90         4            7     38            0                0
#> 91          91         7           11     52            0                0
#> 92          92         3            3     53            0                1
#> 93          93         0            1     92            0                1
#>    vocals_mackaye vocals_lally duration_seconds
#> 1               1            0              401
#> 2               0            0              353
#> 3               0            0              345
#> 4               0            1              327
#> 5               0            1              311
#> 6               1            0              292
#> 7               1            0              289
#> 8               1            0              284
#> 9               0            0              278
#> 10              1            0              278
#> 11              0            0              269
#> 12              0            0              268
#> 13              1            0              267
#> 14              1            0              264
#> 15              0            0              263
#> 16              0            0              261
#> 17              1            0              258
#> 18              1            0              257
#> 19              0            0              256
#> 20              0            0              252
#> 21              1            0              249
#> 22              1            0              242
#> 23              0            0              242
#> 24              0            0              241
#> 25              1            0              239
#> 26              0            0              238
#> 27              0            0              233
#> 28              0            0              232
#> 29              0            1              230
#> 30              0            0              228
#> 31              0            0              226
#> 32              1            0              223
#> 33              1            0              223
#> 34              0            0              222
#> 35              1            0              222
#> 36              1            0              218
#> 37              0            0              216
#> 38              1            0              215
#> 39              0            0              214
#> 40              0            0              212
#> 41              0            0              208
#> 42              1            0              207
#> 43              0            0              207
#> 44              0            0              206
#> 45              0            0              204
#> 46              1            0              201
#> 47              0            0              200
#> 48              0            0              198
#> 49              0            0              196
#> 50              0            0              193
#> 51              1            0              193
#> 52              0            0              191
#> 53              0            0              189
#> 54              0            0              189
#> 55              1            0              188
#> 56              1            0              188
#> 57              0            0              186
#> 58              0            0              185
#> 59              0            0              181
#> 60              1            0              181
#> 61              1            0              179
#> 62              0            0              178
#> 63              1            0              178
#> 64              0            0              177
#> 65              1            0              174
#> 66              1            0              173
#> 67              0            0              173
#> 68              0            0              173
#> 69              1            0              170
#> 70              1            0              170
#> 71              1            0              167
#> 72              0            0              165
#> 73              1            0              162
#> 74              0            0              159
#> 75              1            0              155
#> 76              1            0              154
#> 77              0            0              152
#> 78              0            0              150
#> 79              0            0              150
#> 80              1            0              149
#> 81              0            0              138
#> 82              0            0              137
#> 83              1            0              132
#> 84              1            0              132
#> 85              0            0              130
#> 86              0            0              126
#> 87              0            0              124
#> 88              0            0              121
#> 89              1            0              112
#> 90              1            0              107
#> 91              1            0              105
#> 92              0            0               92
#> 93              0            0              130