Fugazi songs data

Song data from the Fugazi discography pages on Wikipedia. The variables attributing lead vocals are simplifications in some cases where lead vocals were shared.

songvarslookup

Format

dataframe with one row for each song in the Fugazi discography, except those which never appear in the Fugazi Live Series data.

songid

numeric id for each song

releaseid

numeric id in ascending chronological order

release

name of album or EP

track_number

The track number for the song on the release

instrumental

Indicates whether or not the piece is an instrumental

vocals_picciotto

indicates whether or not Guy Picciotto sang lead vocals on this track

vocals_mackaye

indicates whether or not Ian Mackaye sang lead vocals on this track

vocals_lally

indicates whether or not Joe Lally sang lead vocals on this track

duration_seconds

The duration of the song in seconds

Source

https://web.archive.org/web/20201112000517/http://en.wikipedia.org/wiki/Fugazi_discography

Examples

  songvarslookup
#>    songid releaseid             release track_number instrumental
#> 1      83         6 In on the Killtaker            6            0
#> 2      88         9        The Argument            7            0
#> 3      65         8            End Hits            8            0
#> 4      77         9        The Argument            6            0
#> 5      32         7        Red Medicine            8            0
#> 6      44         8            End Hits            7            0
#> 7      16         4            Repeater           11            0
#> 8      15         1              Fugazi            6            0
#> 9      51         6 In on the Killtaker           12            0
#> 10     81        11          First Demo            5            0
#> 11     68         9        The Argument            8            0
#> 12     46         8            End Hits           10            1
#> 13     70         9        The Argument           11            0
#> 14     71         9        The Argument            2            0
#> 15     57         1              Fugazi            7            0
#> 16     79         8            End Hits            9            0
#> 17     73         9        The Argument            9            0
#> 18     61         7        Red Medicine           13            0
#> 19      9         4            Repeater            1            0
#> 20     67         8            End Hits            4            0
#> 21     69         8            End Hits           12            0
#> 22     10         2       Margin Walker            6            0
#> 23     75         9        The Argument           10            0
#> 24     30         6 In on the Killtaker            4            0
#> 25     82         9        The Argument            4            0
#> 26     27         5         Steady Diet            8            0
#> 27     85         9        The Argument            3            0
#> 28      3         4            Repeater            5            0
#> 29     49         8            End Hits            3            0
#> 30     14         6 In on the Killtaker            5            0
#> 31     54         7        Red Medicine            7            0
#> 32     25         6 In on the Killtaker           11            0
#> 33     89        11          First Demo            8            0
#> 34     74         5         Steady Diet            6            1
#> 35     59         8            End Hits           13            0
#> 36     91         5         Steady Diet            9            0
#> 37     31         6 In on the Killtaker            7            1
#> 38     66        10           Furniture            1            0
#> 39     76         7        Red Medicine            3            0
#> 40     29         7        Red Medicine           10            0
#> 41     12         4            Repeater            8            0
#> 42     11         2       Margin Walker            2            0
#> 43     48         5         Steady Diet           10            0
#> 44     60         5         Steady Diet            3            0
#> 45      7         4            Repeater            6            0
#> 46      2         5         Steady Diet            2            0
#> 47     78         7        Red Medicine            9            1
#> 48     50         6 In on the Killtaker           10            0
#> 49     24         7        Red Medicine            1            0
#> 50     39         5         Steady Diet            5            0
#> 51     55         6 In on the Killtaker            3            0
#> 52     22         5         Steady Diet            1            0
#> 53     87         9        The Argument            5            0
#> 54     64        10           Furniture            2            1
#> 55     47         5         Steady Diet            4            0
#> 56     37         7        Red Medicine            4            0
#> 57     86         7        Red Medicine            6            1
#> 58     36         7        Red Medicine            5            0
#> 59     58         3             3 Songs            2            1
#> 60      8         4            Repeater            2            0
#> 61      5         4            Repeater            4            0
#> 62     13         1              Fugazi            5            0
#> 63     40         5         Steady Diet           11            0
#> 64     80         8            End Hits           11            0
#> 65     17         3             3 Songs            1            0
#> 66      1         1              Fugazi            1            0
#> 67     33         1              Fugazi            2            0
#> 68     62         7        Red Medicine           12            0
#> 69     35         2       Margin Walker            3            0
#> 70     18         7        Red Medicine            2            0
#> 71     84        11          First Demo           10            0
#> 72     56         8            End Hits            2            0
#> 73     19         6 In on the Killtaker            1            0
#> 74     34         1              Fugazi            4            0
#> 75     21         1              Fugazi            3            0
#> 76     20         4            Repeater            9            0
#> 77     45         4            Repeater            3            1
#> 78      6         2       Margin Walker            1            0
#> 79     41         6 In on the Killtaker            8            0
#> 80     63         8            End Hits            5            0
#> 81     23         4            Repeater           10            0
#> 82     90         2       Margin Walker            4            0
#> 83      4         5         Steady Diet            7            0
#> 84     42         8            End Hits            1            0
#> 85     43         2       Margin Walker            5            0
#> 86     92        10           Furniture            3            0
#> 87     28         6 In on the Killtaker            2            0
#> 88     72         8            End Hits            6            0
#> 89     26         6 In on the Killtaker            9            0
#> 90     38         4            Repeater            7            0
#> 91     52         7        Red Medicine           11            0
#> 92     53         3             3 Songs            3            0
#>    vocals_picciotto vocals_mackaye vocals_lally duration_seconds
#> 1                 0              1            0              401
#> 2                 1              0            0              353
#> 3                 1              0            0              345
#> 4                 0              0            1              327
#> 5                 0              0            1              311
#> 6                 0              1            0              292
#> 7                 0              1            0              289
#> 8                 0              1            0              284
#> 9                 1              0            0              278
#> 10                0              1            0              278
#> 11                1              0            0              269
#> 12                0              0            0              268
#> 13                0              1            0              267
#> 14                0              1            0              264
#> 15                1              0            0              263
#> 16                1              0            0              261
#> 17                0              1            0              258
#> 18                0              1            0              257
#> 19                1              0            0              256
#> 20                1              0            0              252
#> 21                0              1            0              249
#> 22                0              1            0              242
#> 23                1              0            0              242
#> 24                1              0            0              241
#> 25                0              1            0              239
#> 26                1              0            0              238
#> 27                1              0            0              233
#> 28                1              0            0              232
#> 29                0              0            1              230
#> 30                1              0            0              228
#> 31                1              0            0              226
#> 32                0              1            0              223
#> 33                0              1            0              223
#> 34                0              0            0              222
#> 35                0              1            0              222
#> 36                0              1            0              218
#> 37                0              0            0              216
#> 38                0              1            0              215
#> 39                1              0            0              214
#> 40                1              0            0              212
#> 41                1              0            0              208
#> 42                0              1            0              207
#> 43                1              0            0              207
#> 44                1              0            0              206
#> 45                1              0            0              204
#> 46                0              1            0              201
#> 47                0              0            0              200
#> 48                1              0            0              198
#> 49                1              0            0              196
#> 50                1              0            0              193
#> 51                0              1            0              193
#> 52                1              0            0              191
#> 53                1              0            0              189
#> 54                0              0            0              189
#> 55                0              1            0              188
#> 56                0              1            0              188
#> 57                0              0            0              186
#> 58                1              0            0              185
#> 59                0              0            0              181
#> 60                0              1            0              181
#> 61                0              1            0              179
#> 62                1              0            0              178
#> 63                0              1            0              178
#> 64                1              0            0              177
#> 65                0              1            0              174
#> 66                0              1            0              173
#> 67                1              0            0              173
#> 68                1              0            0              173
#> 69                0              1            0              170
#> 70                0              1            0              170
#> 71                0              1            0              167
#> 72                1              0            0              165
#> 73                0              1            0              162
#> 74                1              0            0              159
#> 75                0              1            0              155
#> 76                0              1            0              154
#> 77                0              0            0              152
#> 78                1              0            0              150
#> 79                1              0            0              150
#> 80                0              1            0              149
#> 81                1              0            0              138
#> 82                1              0            0              137
#> 83                0              1            0              132
#> 84                0              1            0              132
#> 85                1              0            0              130
#> 86                1              0            0              126
#> 87                1              0            0              124
#> 88                1              0            0              121
#> 89                0              1            0              112
#> 90                0              1            0              107
#> 91                0              1            0              105
#> 92                1              0            0               92