Summary

Summary

summary

Format

dataframe with one row for each song in the Fugazi discography, except those which never appear in the Fugazi Live Series data.

rank_rating

The rank of the song in terms of the rating derived from the choice modelling, with the highest-rated song in the first position.

songid

numeric id for each song

song

The name of the song

launchdate

The date on which the song was first performed according to the data

duration_seconds

The duration of the song in seconds

chosen

The number of times the song was performed according to the data

available_rl

The number of shows for which the song was available in the band's repertoire

intensity

The performance intensity is the ratio of chosen/available_rl

rating

Rating on the interval between 0 and 1 where 1 is the highest rating and 0 the lowest.

releaseid

numeric id in ascending chronological order

release

release name

track_number

The track number for the song on the release

instrumental

Indicates whether or not the piece is an instrumental

vocals_picciotto

indicates whether or not Guy Picciotto sang lead vocals on this track

vocals_mackaye

indicates whether or not Ian Mackaye sang lead vocals on this track

vocals_lally

indicates whether or not Joe Lally sang lead vocals on this track

Source

https://www.dischord.com/fugazi_live_series

Examples

  summary
#> # A tibble: 92 × 16
#>    rank_rating songid song       launchdate duration_seconds chosen available_rl
#>          <int>  <dbl> <chr>      <date>                <int>  <dbl>        <dbl>
#>  1           1     18 bed for t… 1994-11-20              170    299          376
#>  2           2      2 reclamati… 1990-05-05              201    594          793
#>  3           3     42 break      1996-08-15              132    170          228
#>  4           4     24 do you li… 1994-11-20              196    272          376
#>  5           5     44 closed ca… 1997-06-18              292    158          198
#>  6           6     71 cashout    2000-09-30              264     59           65
#>  7           7     56 place pos… 1996-08-15              165    133          228
#>  8           8     29 target     1994-08-15              212    261          384
#>  9           9      1 waiting r… 1987-09-03              173    628          895
#> 10          10     49 recap mod… 1997-05-03              230    141          206
#> # … with 82 more rows, and 9 more variables: intensity <dbl>, rating <dbl>,
#> #   releaseid <int>, release <chr>, track_number <int>, instrumental <int>,
#> #   vocals_picciotto <int>, vocals_mackaye <int>, vocals_lally <int>