xray

xray

xray

Format

dataframe with one row for each rendition of each show in the Fugazi Live Series data and each type of units (tracks and minutes)

fls_link

a link to the corresponding page of the Fugazi Live Series site.

year

The year of the show,

tour

The tour that the show belongs to.

date

The date of the show.

units

The units can be either tracks or minutes. The remaining columns on any given row will be in these units

songs

songs that were performed at least twice

released

songs that had been released before the date of the show

unreleased

songs that had not been released before the date of the show

debut

songs that were performed for the first time at this show

farewell

songs that were performed for the last time at this show

incumbent

songs that were neither performed for the first time or the last at this show

other

tracks that are something else, for instance: intro, opening remarks, interlude, encore, outro, interruptions, and one-offs.

...

The remaining columns correspond to specific releases in the band's catalogue.

Source

https://www.dischord.com/fugazi_live_series

Examples

  xray
#> # A tibble: 1,790 × 23
#>    fls_link        year tour  date       units songs relea…¹ unrel…² other debut
#>    <chr>          <dbl> <chr> <date>     <chr> <dbl>   <dbl>   <dbl> <dbl> <dbl>
#>  1 <a href='http…  1987 1987… 1987-09-03 minu…  31.0    0       31.0  0.93     8
#>  2 <a href='http…  1987 1987… 1987-09-26 minu…  30.0    0       30.0  6.32     1
#>  3 <a href='http…  1987 1987… 1987-10-07 minu…  37.0    0       37.0  6.62     1
#>  4 <a href='http…  1987 1987… 1987-10-16 minu…  31.8    0       31.8  2.93     2
#>  5 <a href='http…  1987 1987… 1987-11-25 minu…  34.9    0       34.9  7.98     0
#>  6 <a href='http…  1987 1987… 1987-12-03 minu…  45.7    0       45.7  6.04     2
#>  7 <a href='http…  1987 1987… 1987-12-05 minu…  33.7   -5.55    39.2 12.8      0
#>  8 <a href='http…  1987 1987… 1987-12-28 minu…  52.8    0       52.8  2.93     0
#>  9 <a href='http…  1988 1988… 1988-01-21 minu…  44.6    0       44.6  7.21     0
#> 10 <a href='http…  1988 1988… 1988-01-22 minu…  40.5    0       40.5 10.5      0
#> # … with 1,780 more rows, 13 more variables: farewell <dbl>, incumbent <dbl>,
#> #   fugazi <dbl>, margin_walker <dbl>, three_songs <dbl>, repeater <dbl>,
#> #   steady_diet_of_nothing <dbl>, in_on_the_killtaker <dbl>,
#> #   red_medicine <dbl>, end_hits <dbl>, the_argument <dbl>, furniture <dbl>,
#> #   first_demo <dbl>, and abbreviated variable names ¹​released, ²​unreleased